% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{agglomerate_metacoder}
\alias{agglomerate_metacoder}
\title{Agglomerate Metacoder Objects}
\usage{
agglomerate_metacoder(obj, rank, validated = FALSE)
}
\arguments{
\item{obj}{A Taxmap/metacoder object.}

\item{rank}{The rank that will be agllomerated to.}

\item{validated}{This parameter provides a way to override validation steps.  Use carefully.  Default: FALSE}
}
\value{
A taxmap object that has been agglomerated at the specified rank.
}
\description{
A function similar to the \code{\link[phyloseq:tax_glom]{phyloseq::tax_glom}} function,
that assembles abundance data at a specified rank.  This removes subtaxa and reassigns the
values at the specified rank.
}
\details{
This function helps analyzing taxonomic data at intermediate ranks by agglomerating the observation data
and the taxmap object.
}
\examples{
\dontrun{
if(interactive()){
# This example uses data that are no longer available in the MicrobiomeR package,
# however, they can be easily generated with \\code{\\link{MicrobiomeR}{as_basic_format}}.
library(MicrobiomeR)
phylum_obj <- agglomerate_metacoder(obj = MicrobiomeR::basic_silva, rank = "Phylum")
class_obj <- agglomerate_metacoder(obj = MicrobiomeR::basic_silva, rank = "Class")
order_obj <- agglomerate_metacoder(obj = MicrobiomeR::basic_silva, rank = "Order")
 }
}
}
\seealso{
\code{\link[taxa]{filter_taxa}}

\code{\link[MicrobiomeR]{validate_MicrobiomeR_format}}

Other Advanced Metacoder Filters: \code{\link{cov_filter}},
  \code{\link{otu_prevalence_filter}},
  \code{\link{otu_proportion_filter}},
  \code{\link{taxa_prevalence_filter}}
}
\concept{Advanced Metacoder Filters}
