% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-diversity.R
\name{alpha_diversity_measures}
\alias{alpha_diversity_measures}
\title{Alpha Diversity Measures}
\usage{
alpha_diversity_measures(obj, group = "TreatmentGroup")
}
\arguments{
\item{obj}{An object to be converted to a Taxmap object with \code{\link[MicrobiomeR]{create_taxmap}}.}

\item{group}{The "TreatmentGroup" or similar grouping from your metadata to denote sample groups, Default: 'TreatmentGroup'}
}
\value{
Returns a list of alpha diversity measures with metadata.
}
\description{
This function generates various alpha diversity measures include Shannon, Fisher, Coverage, Gini Simpson, and Inverse Simpson.
}
\examples{
\dontrun{
if (interactive()) {
  library(MicrobiomeR)
  data <- analyzed_silva
  measures <- alpha_diversity_measures(data)
  measures$Shannon
}
}
}
\seealso{
View \code{\link[microbiome]{alpha}},\code{\link[microbiome]{meta}} to learn more about generating diversity measures with the microbiome package.
}
