% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-diversity.R
\name{alpha_diversity_plot}
\alias{alpha_diversity_plot}
\alias{alpha_diversity_plots}
\title{Alpha Diversity Plot}
\usage{
alpha_diversity_plot(
  obj,
  measure = "Shannon",
  group = "TreatmentGroup",
  select_otu_table = NULL,
  title = NULL
)

alpha_diversity_plots(
  obj,
  measures = c("Shannon", "GiniSimpson", "InverseSimpson"),
  group = "TreatmentGroup"
)
}
\arguments{
\item{obj}{An object to be converted to a Taxmap object with \code{\link[MicrobiomeR]{create_taxmap}}.}

\item{measure}{Select an alpha diversity measure such as Shannon, Fisher, Coverage, GiniSimpson, and InverseSimpson, Default: 'Shannon'}

\item{group}{The "TreatmentGroup" or similar grouping or column from your metadata to denote sample groups, Default: 'TreatmentGroup'}

\item{select_otu_table}{DEPRECATED. Choose an otu table to analyze, Default: 'otu_proportions'}

\item{title}{The title of the plot, Default: NULL}

\item{measures}{A list of alpha diversity measures such as Shannon, Fisher, Coverage, GiniSimpson, and InverseSimpson, Default: 'c("Shannon", "GiniSimpson", "InverseSimpson")'}
}
\value{
Returns an alpha diversity plot.

Returns a melted dataframe.
}
\description{
Plot the alpha diversity using a violin plot. \code{alpha_diversity_plots} generates plots for all alpha diversity measures.
}
\details{
Alpha diversity helps to determine the species richness (the number of different species in a sample) or evenness (similar abundance level).
We prefer to use \code{Shannon} as it is better for data generated using the QIIME pipeline.
}
\examples{
\dontrun{
if (interactive()) {
  library(MicrobiomeR)
  data <- analyzed_silva
  plot <- alpha_diversity_plot(obj = data,
                               measure = "Shannon")
  plot
}
}
}
\seealso{
\code{\link{alpha_diversity_measures}}, \code{\link[vegan]{diversity}}, \code{\link[ggpubr]{ggviolin}}

Other Visualizations: 
\code{\link{correlation_data}()},
\code{\link{correlation_plot}()},
\code{\link{correlation_plots}()},
\code{\link{heat_tree_parameters}()},
\code{\link{heat_tree_plots}()},
\code{\link{ordination_plot}()},
\code{\link{plot_limits}()},
\code{\link{save_alpha_diversity_plots}()},
\code{\link{save_correlation_plots}()},
\code{\link{save_heat_tree_plots}()},
\code{\link{save_ordination_plots}()},
\code{\link{save_stacked_barplots}()},
\code{\link{stacked_barplot}()},
\code{\link{stacked_barplots}()},
\code{\link{top_coefficients_barplot}()}

Other Visualizations: 
\code{\link{correlation_data}()},
\code{\link{correlation_plot}()},
\code{\link{correlation_plots}()},
\code{\link{heat_tree_parameters}()},
\code{\link{heat_tree_plots}()},
\code{\link{ordination_plot}()},
\code{\link{plot_limits}()},
\code{\link{save_alpha_diversity_plots}()},
\code{\link{save_correlation_plots}()},
\code{\link{save_heat_tree_plots}()},
\code{\link{save_ordination_plots}()},
\code{\link{save_stacked_barplots}()},
\code{\link{stacked_barplot}()},
\code{\link{stacked_barplots}()},
\code{\link{top_coefficients_barplot}()}
}
\concept{Visualizations}
