% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analyzed_greengenes}
\alias{analyzed_greengenes}
\title{Stress Microbiome "analyzed_format" Data (Greengenes)}
\format{An object of class \code{"Taxmap"} with customized observation tables:
\describe{
\item{otu_abundance}{An abundance table with otu_id, taxon_id, and Sample_* columns.}
\item{otu_annotations}{A taxonomy annotations table with otu_id, taxon_id, and multiple rank columns.}
\item{otu_proportions}{A proportion table derived from the otu_abundance table.}
\item{sample_data}{A metadata table with SampleID, BarcodeSequence, LinkerPrimerSequence,
ForwardFastqFile, ReverseFastqFile, TreatmentGroup, SampleName, Description}
\item{phy_tree}{A phylogenetic tree.}
\item{taxa_abundance}{A taxonomy table interpolated from the otu_abundance table with taxon_id,
and Sample_* columns.}
\item{taxa_proportions}{A taxonomy table derived from the taxa_abundance table.}
\item{statistical_data}{A data table that contains statistical analysis comparing treatment groups.}
\item{stats_tax_data}{A data table that combines taxa and statistical datasets for plotting.}
}}
\source{
\href{https://nephele.niaid.nih.gov/user_guide_pipes/#amplicon_pipes}{Nephele-16S Qiime Pipeline}
}
\usage{
data(analyzed_greengenes)
}
\description{
Data from the Nephele-16S-Qiime-Greengenes microbiome pipeline
in the form of a taxmap object.  This taxmap object has been
formatted into the MicrobiomeR "analyzed_format".
}
\references{
Zhang et al. (2019)
(\href{https://www.ncbi.nlm.nih.gov/pubmed/}{PubMed})
}
\keyword{datasets}
