% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{as_custom_format}
\alias{as_custom_format}
\title{As Custom MicrobiomeR Format}
\usage{
as_custom_format(obj, format, change_name_list = NULL, ...)
}
\arguments{
\item{obj}{An object to be converted to a Taxmap object with \code{\link[MicrobiomeR]{create_taxmap}}.}

\item{format}{The name of the format to convert the object to.}

\item{change_name_list}{A list with names of the tables in the observation data, that have matching
values that are used to change the names of the table.   Default: NULL}

\item{...}{An optional list of parameters to use in the as_*_format function specified}
}
\value{
A Taxmap object that we have tried to format with all of our heart.
}
\description{
A function for formatting Taxmap objects in the MicrobiomeR format.  This function
attempts to give more customization than the as_*_format functions.
}
\details{
This function is meant to be more helpful for customizing the Taxmap object.
}
\seealso{
\code{\link[MicrobiomeR]{create_taxmap}},  \code{\link[MicrobiomeR]{which_format}},  \code{\link[MicrobiomeR]{order_metacoder_data}},  \code{\link[MicrobiomeR]{as_MicrobiomeR_format}}

Other Formatting: \code{\link{MicrobiomeR_Formats}},
  \code{\link{as_MicrobiomeR_format}},
  \code{\link{as_analyzed_format}},
  \code{\link{as_basic_format}},
  \code{\link{as_phyloseq_format}},
  \code{\link{as_raw_format}}, \code{\link{melt_taxmap}},
  \code{\link{treatment_matrix}}
}
\concept{Formatting}
