% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{as_phyloseq_format}
\alias{as_phyloseq_format}
\title{As Phyloseq MicrobiomeR Format}
\usage{
as_phyloseq_format(
  obj,
  otu_table = "otu_abundance",
  tax_data = "otu_annotations",
  sample_data = "sample_data",
  phy_tree = "phy_tree"
)
}
\arguments{
\item{obj}{An object to be converted to a Taxmap object with \code{\link[MicrobiomeR]{create_taxmap}}.}

\item{otu_table}{The name of the observation table with OTU data.  Default: NULL}

\item{tax_data}{The name of the observation table with taxonomic annotations.  Default: NULL}

\item{sample_data}{The name of the observation table with metadata.  Default: NULL}

\item{phy_tree}{The name of the observation data with the phylogenetic tree.  Default: NULL}
}
\value{
A Taxmap object in the phyloseq_format.
}
\description{
Converts the Taxmap object to the phyloseq_format.
}
\details{
See the \link{MicrobiomeR_Formats} documentation.
}
\seealso{
\code{\link[MicrobiomeR]{create_taxmap}},\code{\link[MicrobiomeR]{order_metacoder_data}}

Other Formatting: 
\code{\link{MicrobiomeR_Formats}},
\code{\link{as_MicrobiomeR_format}()},
\code{\link{as_analyzed_format}()},
\code{\link{as_basic_format}()},
\code{\link{as_custom_format}()},
\code{\link{as_raw_format}()},
\code{\link{melt_taxmap}()},
\code{\link{treatment_matrix}()}
}
\concept{Formatting}
