% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{bar_plot}
\alias{bar_plot}
\title{Bar Plot}
\usage{
bar_plot(obj, tax_level = "Phylum", fill = "Phylum",
  xlabel = "Samples", title = NULL, palette_values)
}
\arguments{
\item{obj}{DESCRIPTION.}

\item{tax_level}{DESCRIPTION.}

\item{fill}{DESCRIPTION.}

\item{xlabel}{DESCRIPTION.}

\item{title}{DESCRIPTION.}

\item{palette_values}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
