% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bimodality_coefficient}
\alias{bimodality_coefficient}
\title{FUNCTION_TITLE}
\usage{
bimodality_coefficient(x, finite = TRUE, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{finite}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
