% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{combination_palette}
\alias{combination_palette}
\title{Combine Color Palettes}
\usage{
combination_palette(...)
}
\arguments{
\item{...}{You can use any name for your arguments, but the values must be a named list.
The list can only have 4 named members:
\describe{
\item{palette}{This is a palette function that returns a vector of colors.}
\item{args}{This is another named list used for the palette function parameters.}
\item{range}{This is a range \emph{(1:10)} used to subset the color palette vector.}
\item{rev}{This is a logical \emph{(TRUE/FALSE)} used to reverse the color palette.}
}
You can add as many parameters you want in order to combine as many color palettes
as you want.}
}
\value{
The output of this function is another function (grDevoces::colorRampPalette), which takes
a number to generate an interpolated color palette as a character vector.
}
\description{
This function uses dynamic arguments (...) in order to combine multiple
color palettes together.
}
\details{
This function allows you to combine a varying number of color palettes and gives you
the ability to subset and reverse the palettes that are supplied.
}
\examples{
\dontrun{
if(interactive()){
# Below is the code for the viridis_magma_palette function.
# It's a good example of how to use the combination_palette function.
viridis_magma_palette <- function(viridis_number = 800,
                                  viridis_range = 300:viridis_number,
                                  viridis_rev = TRUE,
                                  magma_number = 500,
                                  magma_range = 0:magma_number,
                                  magma_rev = FALSE,
                                  ...) {
    if (!missing(...)){
        v_args = list(n=viridis_number, ...)
        m_args = list(n=magma_number, ...)
    } else {
        v_args = list(n=viridis_number)
        m_args = list(n=magma_number)
        }
   crp <- combination_palette(viridis =
                                list(palette = viridis::viridis,
                                      args = v_args,
                                      range = viridis_range,
                                      rev = viridis_rev),
                               magma =
                                 list(palette = viridis::magma,
                                      args = m_args,
                                      range = magma_range,
                                      rev = magma_rev)
                               )
 return(crp)
}
 }
}
}
\seealso{
\code{\link[grDevices]{colorRamp}}

Other Color Palettes: 
\code{\link{get_color_palette}()},
\code{\link{scico_palette}()},
\code{\link{viridis_magma_palette}()},
\code{\link{viridis_palette}()}
}
\concept{Color Palettes}
