% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{convert_proportions}
\alias{convert_proportions}
\title{Convert Proportions}
\usage{
convert_proportions(melted_df, tax_level)
}
\arguments{
\item{melted_df}{A "melted" dataframe from the metacoder object's data.}

\item{tax_level}{The taxonomic level, Default: 'Phylum'}
}
\description{
Convert Proportions
}
\seealso{
Other Data Manipulators: 
\code{\link{metacoder_comp_func_1}()},
\code{\link{transformer}()},
\code{\link{transposer}()},
\code{\link{vlookup}()}
}
\concept{Data Manipulators}
