% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-plot.R
\name{correlation_plot}
\alias{correlation_plot}
\title{Correlation Plot}
\usage{
correlation_plot(
  obj,
  primary_rank,
  secondary_rank = TRUE,
  wp_value = 0.05,
  pal_func = NULL,
  trans = "logit"
)
}
\arguments{
\item{obj}{An object to be converted to a Taxmap object with \code{\link[MicrobiomeR]{create_taxmap}}.}

\item{primary_rank}{The primary rank used to label the points.}

\item{secondary_rank}{The secondary rank used to color the points.  Can be an integer specifying
the number of supertaxon ranks above the primary rank or the name of a supertaxon rank.  Default: TRUE}

\item{wp_value}{The Wilcoxian P-Value used to represent significant points.  Default: 0.05}

\item{pal_func}{A palette function that returns grDevices::colorRampPalette.}

\item{trans}{Either the name of a transformation object, or the object itself given to \code{\link[ggplot2]{scale_continuous}}.
Built-in transformations include "asn", "atanh", "boxcox", "exp", "identity", "log", "log10", "log1p",
"log2", "logit", "probability", "probit", "reciprocal", "reverse" and "sqrt".}
}
\value{
A 1:1 correlation plot built with ggplot2.
}
\description{
Create a correlation plot from a metacoder/taxmap object.
}
\details{
Correlation plots help to better explain the heat tree findings.
}
\examples{
\dontrun{
if(interactive()){
# This example uses data that are no longer available in the MicrobiomeR package,
# however, they can be easily generated with \code{\link{MicrobiomeR}{as_analyzed_format}}.
library(MicrobiomeR)
analyzed_silva <- as_MicrobiomeR_format(MicrobiomeR::raw_silva_2, "analyzed_format")
correlation_plot(analyzed_silva, primary_rank = "Class", secondary_rank = "Phylum")
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{create_taxmap}},  \code{\link[MicrobiomeR]{validate_MicrobiomeR_format}},  \code{\link[MicrobiomeR]{correlation_data}},  \code{\link[MicrobiomeR]{plot_limits}},  \code{\link[MicrobiomeR]{get_color_palette}}

\code{\link[ggplot2]{ggplot}},  \code{\link[ggplot2]{aes}},  \code{\link[ggplot2]{geom_polygon}},  \code{\link[ggplot2]{geom_point}},  \code{\link[ggplot2]{labs}},  \code{\link[ggplot2]{scale_continuous}},  \code{\link[ggplot2]{scale_manual}},  \code{\link[ggplot2]{guide_legend}},  \code{\link[ggplot2]{geom_abline}}

\code{\link[ggrepel:geom_text_repel]{geom_label_repel}}

\code{\link[forcats]{fct_reorder}}

Other Visualizations: 
\code{\link{alpha_diversity_plot}()},
\code{\link{correlation_data}()},
\code{\link{correlation_plots}()},
\code{\link{heat_tree_parameters}()},
\code{\link{heat_tree_plots}()},
\code{\link{ordination_plot}()},
\code{\link{plot_limits}()},
\code{\link{save_alpha_diversity_plots}()},
\code{\link{save_correlation_plots}()},
\code{\link{save_heat_tree_plots}()},
\code{\link{save_ordination_plots}()},
\code{\link{save_stacked_barplots}()},
\code{\link{stacked_barplot}()},
\code{\link{stacked_barplots}()},
\code{\link{top_coefficients_barplot}()}
}
\concept{Visualizations}
