% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{cov_filter}
\alias{cov_filter}
\title{Coefficient of Variation Filter}
\usage{
cov_filter(obj, coefficient_of_variation, validated = FALSE)
}
\arguments{
\item{obj}{A Taxmap/metacoder object.}

\item{coefficient_of_variation}{The maximum CoV that an OTU can have.}

\item{validated}{This parameter provides a way to override validation steps.  Use carefully.  Default: FALSE}
}
\value{
Returns a taxmap object that contains otu_ids that have passed the above filter.
}
\description{
This function filters OTUs that have a variance higher than the
specified CoV.
}
\details{
This function helps remove OTUs that have an unusually high variance using the coefficient
of variation.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{validate_MicrobiomeR_format}},\code{\link[MicrobiomeR]{character(0)}},\code{\link[MicrobiomeR]{otu_id_filter}}

Other Metacoder Filters: \code{\link{agglomerate_metacoder}},
  \code{\link{otu_id_filter}},
  \code{\link{otu_prevalence_filter}},
  \code{\link{otu_proportion_filter}},
  \code{\link{sample_filter}},
  \code{\link{taxa_prevalence_filter}},
  \code{\link{taxon_id_filter}}
}
\concept{Metacoder Filters}
