% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bar.R
\name{create_bar}
\alias{create_bar}
\title{Create a per-sample barplot of the top \emph{n} OTUs/ASVs}
\usage{
create_bar(
  ps = NULL,
  df_topn = NULL,
  id = "sample_id",
  n = 15,
  ncol_legend = 3,
  name_legend = "OTU",
  RA = TRUE
)
}
\arguments{
\item{ps}{phyloseq-object (optional)}

\item{df_topn}{topn-data.frame (result of \code{\link{prep_bar}})}

\item{id}{name of the 'id'-column (default: 'sample_id')}

\item{n}{number of OTUs/ASVs to plot}

\item{ncol_legend}{number of columns in the legend}

\item{name_legend}{name of the legend}

\item{RA}{specify whether or not the phyloseq is based on relative abundances (logical). Default: TRUE}
}
\value{
returns a per-sample barplot showing the \emph{n} most abundant OTUs.
}
\description{
Takes either a phyloq-object or an (edited) result of \code{\link{prep_bar}} as an input. And
creates a per-sample barplot.
}
\examples{
data(ps_NP)
ps_NP \%>\%
  phyloseq::subset_samples(visit_name == "w1" & birth_mode == "vag") \%>\%
  create_bar() +
  ggplot2::coord_flip()
}
