% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_taxmap}
\alias{create_taxmap}
\title{Get Taxmap Object}
\usage{
create_taxmap(obj)
}
\arguments{
\item{obj}{An object that contains the data being analyzed.  Can be one of the following:
\describe{
\item{Phyloseq Object}{An object generated from the phyloseq package.}
\item{Taxmap Object}{An object generated from the metacoder or taxa package.}
\item{RData file}{An RData file generated from the base::save function.  Can have an extension of .RData or .rda.}
}}
}
\value{
The output generated is a \strong{taxa::taxmap} object.
}
\description{
A function that handles the conversion of objects to Taxmap (\strong{taxa::taxmap}) objects.
}
\details{
This function is used to convert data to metacoder/taxmap objects for \emph{microbiome} analysis.
This function is used at the beginning of every other function to support multiple types of input
for the obj parameter in those functions. For phyloseq inputs, this function
uses \code{\link[metacoder]{parse_phyloseq}} when possible and falls back to
an internal parser for known upstream metacoder namespace breakage.
}
\examples{
\dontrun{
if(interactive()){
 library(MicrobiomeR)
 phy_obj <- MicrobiomeR::phyloseq_silva_2
 mc_obj <- create_taxmap(phy_obj)
 }
}
}
\seealso{
\code{\link[metacoder]{parse_phyloseq}}

\code{\link[tools]{fileutils}}

Other Data Importers: 
\code{\link{create_phyloseq}()}
}
\concept{Data Importers}
