% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_OTU.R
\name{format_OTU}
\alias{format_OTU}
\title{Function to reformat OTU names for plotting.}
\usage{
format_OTU(OTU_names, short = F, parse = F)
}
\arguments{
\item{OTU_names}{vector of OTU-names in the form of <genus>\emph{<species>}<numerical_id>, e.g.
'Dolosigranulum_pigrum_3'}

\item{short}{whether or not the short name (dropping the species-name if available) should be
returned (logical). Default: FALSE}

\item{parse}{option to use in conjunction with \code{\link[ggrepel]{geom_text_repel}} enabling
italicized/formatted OTU names in for example ordination plots (logical). Default: FALSE}
}
\value{
returns a vector of properly formatted OTU names.
}
\description{
Italicize the taxonomical annotations and place the numerical identifier in brackets. By default,
returns names formatted to work in conjunction with \code{\link[ggtext]{element_markdown}}. For example
'Dolosigranulum_pigrum_3' will be formatted into '\emph{Dolosigranulum pigrum} (3)' in the figure.
}
\examples{
phyloseq::taxa_names(ps_NP)[1:10] \%>\% format_OTU()

}
