% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-diversity.R
\name{get_alpha_diversity}
\alias{get_alpha_diversity}
\title{Get Alpha Diversity}
\usage{
get_alpha_diversity(obj)
}
\arguments{
\item{obj}{A phyloseq or metacoder object.}
}
\value{
Returns a list an alpha diversity object.
}
\description{
This function generates various alpha diversity measures include Shannon, Fisher, Coverage, Gini Simpson, and Inverse Simpson.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[microbiome]{diversities}},\code{\link[microbiome]{meta}}
}
