% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-diversity.R
\name{get_alpha_diversity_measures}
\alias{get_alpha_diversity_measures}
\title{Get Alpha Diversity}
\usage{
get_alpha_diversity_measures(obj)
}
\arguments{
\item{obj}{An object to be converted to a metacoder object with \code{\link[MicrobiomeR]{object_handler}}.}
}
\value{
Returns a list of alpha diversity measures with metadata.
}
\description{
This function generates various alpha diversity measures include Shannon, Fisher, Coverage, Gini Simpson, and Inverse Simpson.
}
\examples{
\dontrun{
if (interactive()) {
  library(MicrobiomeR)
  data <- analyzed_silva
  measures <- get_alpha_diversity(data)
  measures$Shannon
}
}
}
\seealso{
\code{\link[microbiome]{diversities}},\code{\link[microbiome]{meta}}
}
