% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Get Color palette}
\usage{
get_color_palette(pal_func = viridis_magma_palette, color_no = 20,
  display = TRUE, ...)
}
\arguments{
\item{pal_func}{A function that returns the output from grDevoces::colorRampPalette. Default: viridis_magma_palette}

\item{color_no}{The number of colors in the palette. Default: 20}

\item{display}{Boolean for displaying a pie chart of the palette. Default: TRUE}

\item{...}{Parameters for the pal_func.}
}
\value{
Returns a color palette in the form of a character vector.
}
\description{
Get a color palette with a specific number of colors.
}
\details{
This function is meant to be a plugin style function for user created palettes.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other Color Palettes: \code{\link{combination_palette}},
  \code{\link{scico_palette}},
  \code{\link{viridis_magma_palette}},
  \code{\link{viridis_palette}}
}
\concept{Color Palettes}
