% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plot.R
\name{get_correlation_plot}
\alias{get_correlation_plot}
\title{FUNCTION_TITLE}
\usage{
get_correlation_plot(label_rank, phyloseq_object, super_taxa = 1,
  wp_value = 0.05, plotly = FALSE, excel_export = FALSE,
  plotly_export = FALSE, ...)
}
\arguments{
\item{label_rank}{PARAM_DESCRIPTION}

\item{phyloseq_object}{PARAM_DESCRIPTION}

\item{super_taxa}{PARAM_DESCRIPTION, Default: 1}

\item{wp_value}{PARAM_DESCRIPTION, Default: 0.05}

\item{plotly}{PARAM_DESCRIPTION, Default: FALSE}

\item{excel_export}{PARAM_DESCRIPTION, Default: FALSE}

\item{plotly_export}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{tax_glom}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{filter}},\code{\link[dplyr]{arrange}}
\code{\link[stringr]{str_detect}}
\code{\link[plotly]{ggplotly}}
\code{\link[DT]{dataTableOutput}}
}
\concept{FAMILY_TITLE}
