% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{get_distance_methods}
\alias{get_distance_methods}
\title{Get Distance Methods}
\usage{
get_distance_methods(tree_methods = TRUE)
}
\arguments{
\item{tree_methods}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
The function for manipulating phyloseqs native distanceMethodList
}
