% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-tree.R
\name{get_heat_tree_parameters}
\alias{get_heat_tree_parameters}
\title{Get Heat Tree Parameters}
\usage{
get_heat_tree_parameters(obj, title, ...)
}
\arguments{
\item{obj}{A metacoder object.}

\item{title}{The title used in the heat_tree plot.}

\item{...}{Any of the heat tree parameters list below can be used to change the way the heat_tree
output is displayed.  However, this function acts as a default list of parameters.  The memebers
of the default list will be overridden by the dot parameters.}
}
\value{
A list used with do.call and the metacoder::heat_tree function.
}
\description{
This function get's the parameters used for the get_heat_tree_plots function.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[metacoder]{heat_tree}}

\code{\link[taxa]{n_obs}},\code{\link[taxa]{taxon_names}}
\code{\link[purrr]{list_modify}}

Other Visualizations: \code{\link{correlation_plot}},
  \code{\link{get_correlation_plots}},
  \code{\link{get_heat_tree_plots}},
  \code{\link{get_plot_limits}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}}
}
\concept{Visualizations}
