% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-tree.R
\name{get_heat_tree_plots}
\alias{get_heat_tree_plots}
\title{Get Heat Tree Plots}
\usage{
get_heat_tree_plots(obj, rank_list = NULL, ...)
}
\arguments{
\item{obj}{An object to be converted to a metacoder object with \code{\link[MicrobiomeR]{object_handler}}.}

\item{rank_list}{A vector of ranks used to generate heat_trees.  Default: NULL}

\item{...}{Any of the \code{\link[metacoder]{heat_tree}} parameters can be used to change the way the heat_tree
output is displayed.  Please see the \code{\link[MicrobiomeR]{get_heat_tree_parameters}} documentation
for further explanation.}
}
\value{
A list of heat_tree plots.
}
\description{
A function for getting multiple heat_tree plots per rank.
}
\examples{
\dontrun{
if(interactive()){
# This example uses data that are no longer available in the MicrobiomeR package,
# however, they can be easily generated with \\code{\\link{MicrobiomeR}{as_analyzed_format}}.
library(MicrobiomeR)
analyzed_silva <- as_MicrobiomeR_format(MicrobiomeR::raw_silva_2, "analyzed_format")
h_trees <- get_heat_tree_plots(analyzed_silva, rank_list = c("Phylum", "Class"))
h_trees$Class
 }
}
}
\seealso{
\code{\link[metacoder]{heat_tree}}

\code{\link[MicrobiomeR]{object_handler}},  \code{\link[MicrobiomeR]{validate_MicrobiomeR_format}},  \code{\link[MicrobiomeR]{get_heat_tree_parameters}}

\code{\link[taxa]{filter_obs}}

\code{\link[crayon]{crayon}}

\code{\link[ggplot2]{theme}},  \code{\link[ggplot2:element]{margin}},  \code{\link[ggplot2]{labs}}

Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_plot}},
  \code{\link{get_correlation_data}},
  \code{\link{get_correlation_plots}},
  \code{\link{get_heat_tree_parameters}},
  \code{\link{get_plot_limits}},
  \code{\link{save_barplot}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{stacked_barplot}}
}
\concept{Visualizations}
