% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{get_metacoder_obj}
\alias{get_metacoder_obj}
\title{Get Metacoder Object}
\usage{
get_metacoder_obj(phyloseq_object = NULL, rdata_file = NULL,
  filter_flag = FALSE, ambiguous_filter = c("uncultured", "Unassigned",
  "Ambiguous"), kingdom_filter = c("Bacteria"),
  workflow_func = metacoder_workflow_1,
  comp_func = metacoder_comp_func_1)
}
\arguments{
\item{phyloseq_object}{A phyloseq object.  Default: NULL}

\item{rdata_file}{A .Rdata file.  Default: NULL}

\item{filter_flag}{A flag that will determine if the metacoder object gets filtered.  Default: FALSE}

\item{ambiguous_filter}{A vector denoting ambiguous taxonomies to filter.  Default: c("uncultured", "Unassigned", "Ambiguous")}

\item{kingdom_filter}{A character vector denoting which Kingodm's to keep. Default: c("Bacteria")}

\item{workflow_func}{A function that takes a metacoder object and a comparison function (See \code{\link{metacoder_workflow_1}})}

\item{comp_func}{A function used for metacoders compare_groups function.  (See \code{\link{metacoder_comp_func_1}})}
}
\value{
Returns a metacoder object to be used for generating a heat tree.
}
\description{
Create a metacoder object using a phyloseq object or rdata file.
}
\details{
Creates a metacoder object that by default compares treatment groups for downstream analysis.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[metacoder]{parse_phyloseq}},\code{\link[metacoder]{calc_obs_props}},\code{\link[metacoder]{calc_taxon_abund}},\code{\link[metacoder]{compare_groups}}
}
\concept{Data Importer}
