% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{get_phyloseq_obj}
\alias{get_phyloseq_obj}
\title{Get Phyloseq Object}
\usage{
get_phyloseq_obj(b_file = NULL, t_file = NULL, md_file = NULL,
  parse_func = NULL, rdata_file = NULL)
}
\arguments{
\item{b_file}{DESCRIPTION.}

\item{t_file}{DESCRIPTION.}

\item{md_file}{DESCRIPTION.}

\item{parse_func}{DESCRIPTION.}

\item{rdata_file}{DESCRIPTION.}
}
\value{
A phyloseq object
}
\description{
Import or Create a phyloseq object
}
