% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topn.R
\name{get_topn}
\alias{get_topn}
\title{Extract the top \emph{n} OTUs from a phyloseq-object}
\usage{
get_topn(ps, n = 15, residuals = TRUE)
}
\arguments{
\item{ps}{phyloseq-object.}

\item{n}{number of OTUs to extract; ranking based on mean (relative) abundance.}

\item{residuals}{include residuals (logical).}
}
\value{
Returns a phyloseq-object only containing the top \emph{n} OTUs.
This function is mainly used in the context of plotting functions.
}
\description{
Extract the top \emph{n} OTUs from a phyloseq-object
}
\examples{
data(enterotype)
get_topn(enterotype)
}
