% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{is_analyzed_format}
\alias{is_analyzed_format}
\title{Is Analyzed MicrobiomeR Format}
\usage{
is_analyzed_format(obj)
}
\arguments{
\item{obj}{A Taxmap/metacoder object.}
}
\value{
A logical (TRUE/FALSE).
}
\description{
This function returns a logical based on weather or not the object is in the analyzed_format.
}
\details{
The "analyzed_format" is Level 3. in the \link{MicrobiomeR_Formats} hierarchy.
}
\seealso{
Other Validation: \code{\link{is_basic_format}},
  \code{\link{is_phyloseq_format}},
  \code{\link{is_raw_format}},
  \code{\link{order_metacoder_data}},
  \code{\link{validate_MicrobiomeR_format}},
  \code{\link{which_format}}
}
\concept{Validation}
