% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{is_basic_format}
\alias{is_basic_format}
\title{Is Basic MicrobiomeR Format}
\usage{
is_basic_format(obj)
}
\arguments{
\item{obj}{A Taxmap object.}
}
\value{
A logical (TRUE/FALSE).
}
\description{
This function returns a logical based on weather or not the object is in the basic_format.
}
\details{
The "basic_format" is Level 2. in the \link{MicrobiomeR_Formats} hierarchy.
}
\examples{
\dontrun{
if(interactive()){
library(MicrobiomeR)
if(is_basic_format(MicrobiomeR::raw_silva_2)){
    print("It's basic!")
} else {
    fmt <- which_format(MicrobiomeR::raw_silva_2)
    print("It's not analyzed!")
    print(sprintf("It's \%s!", fmt))
}
 }
}
}
\seealso{
Other Validation: 
\code{\link{is_analyzed_format}()},
\code{\link{is_phyloseq_format}()},
\code{\link{is_raw_format}()},
\code{\link{order_metacoder_data}()},
\code{\link{validate_MicrobiomeR_format}()},
\code{\link{which_format}()}
}
\concept{Validation}
