% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{is_phyloseq_format}
\alias{is_phyloseq_format}
\title{Is Phyloseq MicrobiomeR Format}
\usage{
is_phyloseq_format(obj)
}
\arguments{
\item{obj}{A Taxmap/metacoder object.}
}
\value{
A logical (TRUE/FALSE).
}
\description{
This function returns a logical based on weather or not the object is in the phyloseq_format.
}
\details{
The "phyloseq_format" is Level 0. in the \link{MicrobiomeR_Formats} hierarchy.
}
\examples{
\dontrun{
if(interactive()){
library(MicrobiomeR)
if(is_analyzed_format(MicrobiomeR::raw_silva)){
    print("It's phyloseq format!")
} else {
    fmt <- which_format(MicrobiomeR::raw_silva)
    print("It's not phyloseq fomrat!")
    print(sprintf("It's \%s!", fmt))
}
 }
}
}
\seealso{
Other Validation: \code{\link{is_analyzed_format}},
  \code{\link{is_basic_format}},
  \code{\link{is_raw_format}},
  \code{\link{order_metacoder_data}},
  \code{\link{validate_MicrobiomeR_format}},
  \code{\link{which_format}}
}
\concept{Validation}
