% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{is_raw_format}
\alias{is_raw_format}
\title{Is Raw MicrobiomeR format}
\usage{
is_raw_format(obj)
}
\arguments{
\item{obj}{A Taxmap/metacoder object.}
}
\value{
A logical (TRUE/FALSE).
}
\description{
This function returns a logical based on weather or not the object is in the raw_format.
}
\details{
The "raw_format" is the most basic of the \link{MicrobiomeR_Formats} other than the "phyloseq_format".
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{which_format}}

Other Formatting and Validation: \code{\link{object_handler}},
  \code{\link{which_format}}
}
\concept{Formatting and Validation}
