% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log10_px.R
\name{log10_px}
\alias{log10_px}
\title{log10 pseudocount conversion}
\usage{
log10_px(ps, pseudocount = 1)
}
\arguments{
\item{ps}{phyloseq-object}

\item{pseudocount}{add a pseudocount (default = 1); prevents conversion-problems with zero.}
}
\value{
converted phyloseq-object
}
\description{
log10 pseudocount conversion
}
\examples{
data(ps_NP)
log10_px(ps_NP, 0.00001)
}
