% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_color_scheme.R
\name{make_color_scheme}
\alias{make_color_scheme}
\title{Expands colors of any brewer.pal-scheme to \emph{n} colors}
\usage{
make_color_scheme(name, n)
}
\arguments{
\item{name}{name of the color scheme, e.g. 'Paired', 'Spectral' or 'BrBG'.}

\item{n}{number of colors needed.}
}
\value{
returns a vector of colors which can be used in various plots/plotting functions.
}
\description{
Expands colors of any brewer.pal-scheme to \emph{n} colors
}
\examples{
make_color_scheme("Paired", 15)

}
\seealso{
For scheme-options see \code{\link[RColorBrewer:RColorBrewer]{RColorBrewer::brewer.pal.info()}}.
}
