% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{melt_metacoder}
\alias{melt_metacoder}
\title{Melt Metacoder Object}
\usage{
melt_metacoder(obj)
}
\arguments{
\item{obj}{An object to be converted to a metacoder object with \code{\link[MicrobiomeR]{object_handler}}.}
}
\value{
Returns a melted dataframe.
}
\description{
Melt the metacoder or phyloseq tables into a dataframe.
}
