% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{melt_metacoder_obj}
\alias{melt_metacoder_obj}
\title{Melt Metacoder Object}
\usage{
melt_metacoder_obj(obj)
}
\arguments{
\item{obj}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Melt the metacoder tables into a dataframe.
}
