% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{melt_taxmap}
\alias{melt_taxmap}
\title{Melt Taxmap}
\usage{
melt_taxmap(obj)
}
\arguments{
\item{obj}{An object to be converted to a taxmap object with \code{\link[MicrobiomeR]{create_taxmap}}.}
}
\description{
Melt Taxmap
}
\seealso{
Other Formatting: 
\code{\link{MicrobiomeR_Formats}},
\code{\link{as_MicrobiomeR_format}()},
\code{\link{as_analyzed_format}()},
\code{\link{as_basic_format}()},
\code{\link{as_custom_format}()},
\code{\link{as_phyloseq_format}()},
\code{\link{as_raw_format}()},
\code{\link{treatment_matrix}()}
}
\concept{Formatting}
