% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_by_group}
\alias{merge_by_group}
\title{Merge By Group}
\usage{
merge_by_group(phyloseq_object, group_name)
}
\arguments{
\item{phyloseq_object}{A phyloseq object.}

\item{group_name}{The name of the group to merge samples by.}
}
\value{
A merged phyloseq object.
}
\description{
Merge samples by group such as control or experimental.
}
