% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_to_df.R
\name{meta_to_df}
\alias{meta_to_df}
\title{Convert phyloseq metadata (sample_data) to dataframe.}
\usage{
meta_to_df(ps, sample_name = "sample_id")
}
\arguments{
\item{ps}{phyloseq-object}

\item{sample_name}{name of the column containing samples (default: 'sample_id')}
}
\value{
A dataframe with metadata contained in phyloseq.
}
\description{
Convert phyloseq metadata (sample_data) to dataframe.
}
\examples{
data(enterotype)
meta_to_df(enterotype)
}
