% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{metacoder_initial_filter}
\alias{metacoder_initial_filter}
\title{Metacoder Inital Filter}
\usage{
metacoder_initial_filter(metacoder_obj, ambiguous_filter, kingdom_filter)
}
\arguments{
\item{metacoder_obj}{A metacoder object.}

\item{ambiguous_filter}{A vector of ambiguous taxonomy annotations to filter out.}

\item{kingdom_filter}{A vector of Kingodm level taxonomies to keep.}
}
\value{
Returns a filtered metacoder object.
}
\description{
A function for filtering ambiguous taxonomy annotations and unwanted Kindgom level taxonomies.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[taxa]{filter_taxa}}
\code{\link[stringr]{str_detect}}

Other Filters: \code{\link{preprocess_phyloseq}},
  \code{\link{remove_ambiguous_taxa}}
}
\concept{Filters}
