% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{metacoder_workflow_1}
\alias{metacoder_workflow_1}
\title{Metacoder Workflow Function #1}
\usage{
metacoder_workflow_1(metacoder_obj, func = metacoder_comp_func_1)
}
\arguments{
\item{metacoder_obj}{A taxmap generated by metacoder.}

\item{func}{A comparison function for use with \code{\link[metacoder]{compare_groups}} Default: metacoder_comp_func_1}
}
\value{
Returns a altered metacoder object.
}
\description{
A function for analyzing or manipulating metacoder taxmap objects.
}
\details{
Takes a metacoder object and a function for manipulating metacoder objects.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[metacoder]{calc_obs_props}},\code{\link[metacoder]{calc_taxon_abund}},\code{\link[metacoder]{compare_groups}}

Other Data Manipulators: \code{\link{metacoder_comp_func_1}},
  \code{\link{phyloseq_to_dataframe}},
  \code{\link{phyloseq_to_excel}},
  \code{\link{phyloseq_to_stats_dataframe}},
  \code{\link{phyloseq_to_tibble}}, \code{\link{vlookup}}
}
\concept{Data Manipulators}
