% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{MicrobiomeR_Formats}
\alias{MicrobiomeR_Formats}
\title{The MicrobiomeR Formats}
\description{
The formats are listed in order from lowest to highest level.
To get a taxmap object into the higher formats they must be in one of the lower formats.
It is also inherently important to understand the \link{MicrobiomeR_Workflow}, which describes some of the important
dependencies used to create the MicrobiomeR formats listed below:
\describe{
\item{phyloseq_format}{Level 0.  The "phyloseq_format" is the format given to a \emph{taxmap} object that has just
been converted from a \emph{phyloseq} object using \emph{metacoder} via \code{\link[MicrobiomeR]{create_phyloseq}}.  A taxmap object identified by MicrobiomeR
to be in this format will contain "otu_table", "tax_data", "sample_data" and "phy_tree" data in the observation
tables (\strong{metacoder_obj$data}), which represents the matching \emph{phyloseq::otu_table}, \emph{phyloseq::tax_table},
\emph{phyloseq::sample_data}, and \emph{phyloseq::phy_tree} data in the original phyloseq object.}
\item{raw_format}{Level 1.  The "raw_format" is the format given to a \emph{taxmap} object that has been
processed with \code{\link[MicrobiomeR]{as_raw_format}}.  It contains the "otu_abundance" and "otu_annotations" data in the observation
tables, which are just name conversions of the \emph{otu_table} and \emph{tax_data} table from the "phyloseq_format".}
\item{basic_format}{Level 2.  The "basic_format" is the format given to a \emph{taxmap} object that has been
processed with \code{\link[MicrobiomeR]{as_basic_format}}.  It contains the tables available in the "raw_format" along with
"taxa_abundance", "otu_proportions", and "taxa_proportions" tables.}
\item{analyzed_format}{Level 3.  The "analyzed_format" is the format given to a \emph{taxmap} object that has been
processed with \code{\link[MicrobiomeR]{as_analyzed_format}}.  It contains the tables available in the "basic_format" along
with "statistical_data" and "stats_tax_data" tables.}
}
}
\seealso{
Other Formatting: \code{\link{as_MicrobiomeR_format}},
  \code{\link{as_analyzed_format}},
  \code{\link{as_basic_format}},
  \code{\link{as_custom_format}},
  \code{\link{as_phyloseq_format}},
  \code{\link{as_raw_format}},
  \code{\link{melt_metacoder}},
  \code{\link{treatment_matrix}}
}
\concept{Formatting}
