% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mkdir}
\alias{mkdir}
\title{Make a Directory}
\usage{
mkdir(dirname, path = NULL)
}
\arguments{
\item{dirname}{The name of the directory to create}

\item{path}{The path to create the directory in.}
}
\value{
Returns the absolute path of the directory.
}
\description{
This is a custom function for creating a directory.
}
\details{
This function will create a directory only if it doesn't exist and then return the path of that directory.
}
