% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{object_handler}
\alias{object_handler}
\title{Object handler}
\usage{
object_handler(obj)
}
\arguments{
\item{obj}{An object that contains the data being analyzed.  Can be one of the following:
\describe{
\item{Phyloseq Object}{An object generated from the phyloseq package.}
\item{Taxmap Object}{An object generated from the metacoder or taxa package.}
\item{RData file}{An RData file generated from the base::save function.  Can have an extension of .RData or .rda.}
}}
}
\value{
The output generated is a taxmap object.
}
\description{
A function that handles the conversion of objects to metacoder (taxa::taxmap) objects.
}
\details{
This function is used to convert data to metacoder/taxmap objects for microbiome analysis.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[metacoder]{parse_phyloseq}}
\code{\link[tools]{fileutils}}

Other Formatting and Validation: \code{\link{is_raw_format}},
  \code{\link{which_format}}
}
\concept{Formatting and Validation}
