% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{order_metacoder_data}
\alias{order_metacoder_data}
\title{Order Metacoder Observation Data}
\usage{
order_metacoder_data(obj)
}
\arguments{
\item{obj}{An object to be converted to a Taxmap object with \code{\link[MicrobiomeR]{create_taxmap}}.}
}
\value{
A taxmap object with observation data in the proper order for downstream analysis.
}
\description{
A function for changing the order of the observation data in a Taxmap object
}
\details{
Changes the order of the observation tables in \strong{metacoder_object$data} to
otu_abundance, otu_annotations, otu_proportions, sample_data, phy_tree, taxa_abundance,
taxa_proportions, statistical_data, and stats_tax_data respectively.
}
\seealso{
Other Validation: 
\code{\link{is_analyzed_format}()},
\code{\link{is_basic_format}()},
\code{\link{is_phyloseq_format}()},
\code{\link{is_raw_format}()},
\code{\link{validate_MicrobiomeR_format}()},
\code{\link{which_format}()}
}
\concept{Validation}
