% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{ordination_plots}
\alias{ordination_plots}
\title{Ordination Plots}
\usage{
ordination_plots(
  obj,
  methods = c("PCoA", "NMDS"),
  distances = c("wunifrac", "unifrac", "bray"),
  color = "TreatmentGroup",
  select_otu_table = "otu_proportions"
)
}
\arguments{
\item{obj}{An object to be converted to a metacoder object with \code{\link[MicrobiomeR]{create_taxmap}}.}

\item{methods}{A list of ordination methods, Default: 'c("PCoA", "NMDS")'}

\item{distances}{A list of distance methods, Default: 'c("wunifrac", "unifrac", "bray")'}

\item{color}{Choose the group or factor of which colors will be mapped to, Default: 'TreatmentGroup'}

\item{select_otu_table}{The data table to use in the observation data.  Default:  "otu_proportions"}
}
\value{
Returns a melted dataframe.
}
\description{
Generate plots for a list of ordination methods and distances.
}
\seealso{
Other Visualizations: 
\code{\link{alpha_diversity_plot}()},
\code{\link{correlation_data}()},
\code{\link{correlation_plots}()},
\code{\link{correlation_plot}()},
\code{\link{heat_tree_parameters}()},
\code{\link{heat_tree_plots}()},
\code{\link{ordination_plot}()},
\code{\link{plot_limits}()},
\code{\link{save_alpha_diversity_plots}()},
\code{\link{save_correlation_plots}()},
\code{\link{save_heat_tree_plots}()},
\code{\link{save_ordination_plots}()},
\code{\link{save_stacked_barplots}()},
\code{\link{stacked_barplots}()},
\code{\link{stacked_barplot}()},
\code{\link{top_coefficients_barplot}()}
}
\concept{Visualizations}
