% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{otu_id_filter}
\alias{otu_id_filter}
\title{Filter OTU Ids from Taxmap Objects}
\usage{
otu_id_filter(
  obj,
  .f_transform = NULL,
  .f_filter = NULL,
  .f_condition = NULL,
  validated = FALSE,
  ...
)
}
\arguments{
\item{obj}{A Taxmap object.}

\item{.f_transform}{A function used for transforming the data.  Default: NULL}

\item{.f_filter}{A function used for summarising the data like 'sum' or 'mean'.  Default: NULL}

\item{.f_condition}{A function that takes the summarised data and applied a condition like x > 10000.  Default: NULL}

\item{validated}{This parameter provides a way to override validation steps.  Use carefully.  Default: FALSE}

\item{...}{An optional list of parameters to use in the .f_filter function specified}
}
\value{
Returns a taxmap object with otu_ids that pass the filters.
}
\description{
This function provides a flexible way to filter unwanted otu_ids from the taxmap object and from the
observations of a MicrobiomeR formatted object.
}
\details{
Get the otu_ids to keep by using purr and the user supplied transform and filter + condition formulas.
The purr package allows the use of anonymous functions as described in the link below:

\url{https://jennybc.github.io/purrr-tutorial/ls03_map-function-syntax.html#anonymous_function,_formula}
}
\examples{
\dontrun{
if(interactive()){
# Below is the code for the otu_prevelence_filter function.
# It's a good example of how to use the otu_id_filter function.
library(MicrobiomeR)
otu_proportion_filter <- function(obj, otu_percentage = 0.00005, validated = FALSE) {
    mo_clone <- obj$clone()
    mo_clone <- validate_MicrobiomeR_format(obj = mo_clone,
                                            valid_formats = c("raw_format", "basic_format"),
                                            force_format = TRUE,
                                            validated = validated,
                                            min_or_max = min)
    # Filter OTU ids
    mo_clone <- otu_id_filter(obj = mo_clone,
                              .f_transform = ~./sum(.),
                              .f_filter = ~mean(.),
                              .f_condition = ~.> otu_percentage)
return(mo_clone)
}
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{validate_MicrobiomeR_format}},  \code{\link[MicrobiomeR]{transposer}},  \code{\link[MicrobiomeR]{transformer}}

\code{\link[metacoder]{get_dataset}},  \code{\link[metacoder]{filter_taxa}}

\code{\link[dplyr]{select_all}}

\code{\link[purrr]{map}},  \code{\link[purrr]{keep}}

Other Basic Metacoder Filters: 
\code{\link{sample_id_filter}()},
\code{\link{taxon_id_filter}()}
}
\concept{Basic Metacoder Filters}
