% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{otu_prevalence_filter}
\alias{otu_prevalence_filter}
\title{OTU Prevalence Filter}
\usage{
otu_prevalence_filter(
  obj,
  minimum_abundance = 5,
  rel_sample_percentage = 0.5,
  validated = FALSE
)
}
\arguments{
\item{obj}{A Taxmap object.}

\item{minimum_abundance}{The minimum abundance needed per observation per sample.  Default: 5}

\item{rel_sample_percentage}{The percentage of samples per observation that meet the minimum abundance.  Default: 0.5}

\item{validated}{This parameter provides a way to override validation steps.  Use carefully.  Default: FALSE}
}
\value{
Returns a taxmap object that contains taxon_ids that have passed the above filter.
}
\description{
This function filters observations by their prevalence across samples.
}
\details{
The otu_prevalence_filter filters taxon_ids that do not appear more than a certain amount of times (minimum abundance) in a certain percentage of
samples (rel_sample_percentage). The \href{http://web.stanford.edu/class/bios221/MicrobiomeWorkflowII.html#filtering}{phyloseq workflow} calls for a minimum abundance of 5 across \%50 of the samples.
This filtering method is considered unsupervised, because it solely relies on the data in this experiment (OTU ids).
}
\examples{
\dontrun{
if(interactive()){
library(MicrobiomeR)
library(metacoder)
library(taxa)

# Convert Phyloseq object to taxmap object
metacoder_obj <- as_MicrobiomeR_format(obj = phyloseq_obj, format = "raw_format")

# Remove Archaea from the taxmap object
metacoder_obj <- filter_taxa(
  obj = metacoder_obj,
  taxon_names == "Archaea",
  subtaxa = TRUE,
  invert = TRUE)

# Ambiguous Annotation Filter - Remove taxonomies with ambiguous names
metacoder_obj <- filter_ambiguous_taxa(metacoder_obj, subtaxa = TRUE)

# Low Sample Filter - Remove the low samples
metacoder_obj <- sample_id_filter(obj          = metacoder_obj,
                               .f_filter    = ~sum(.),
                               .f_condition = ~.>= 20, validated = TRUE)

# Master Threshold Filter - Add the otu_proportions table and then filter OTUs based on min \%
metacoder_obj <- otu_proportion_filter(
    obj = metacoder_obj,
    otu_percentage = 0.00001
    )

# OTU prevalence filter
metacoder_obj <- otu_prevalence_filter(obj = metacoder_obj, validated = TRUE)
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{validate_MicrobiomeR_format}}

\code{\link[metacoder]{calc_prop_samples}}

\code{\link[dplyr]{filter}}

\code{\link[metacoder]{filter_taxa}}

Other Advanced Metacoder Filters: 
\code{\link{agglomerate_taxmap}()},
\code{\link{cov_filter}()},
\code{\link{otu_proportion_filter}()},
\code{\link{taxa_prevalence_filter}()}
}
\concept{Advanced Metacoder Filters}
