% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{otu_proportion_filter}
\alias{otu_proportion_filter}
\title{OTU Proportion FIlter}
\usage{
otu_proportion_filter(obj, otu_percentage = 5e-05, validated = FALSE)
}
\arguments{
\item{obj}{A Taxmap/metacoder object.}

\item{otu_percentage}{The minimum percentage used to compare against the proportional OTU mean.  Default: 5e-05}

\item{validated}{This parameter provides a way to override validation steps.  Use carefully.  Default: FALSE}
}
\value{
Returns a taxmap object that contains otu_ids that have passed the above filter.
}
\description{
This function filters OTU values from the observation data and the taxmap object
based on a minimum proportional mean across samples per OTU.
}
\details{
This type of filtering is used to remove OTUs that do not have a specified mean proportion.
This function must be used conservatively, hence why the default otu_percentage is so low.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{validate_MicrobiomeR_format}},\code{\link[MicrobiomeR]{character(0)}}

Other Metacoder Filters: \code{\link{agglomerate_metacoder}},
  \code{\link{cov_filter}}, \code{\link{otu_id_filter}},
  \code{\link{otu_prevalence_filter}},
  \code{\link{sample_filter}},
  \code{\link{taxa_prevalence_filter}},
  \code{\link{taxon_id_filter}}
}
\concept{Metacoder Filters}
