% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_tab_to_df.R
\name{otu_tab_to_df}
\alias{otu_tab_to_df}
\title{Convert phyloseq OTU-table (otu_table) to dataframe.}
\usage{
otu_tab_to_df(ps, sample_name = "sample_id")
}
\arguments{
\item{ps}{phyloseq-object}

\item{sample_name}{name of the column containing samples (default: 'sample_id')}
}
\value{
A dataframe with OTU-data contained in phyloseq.
}
\description{
Convert phyloseq OTU-table (otu_table) to dataframe.
}
\examples{
data(ps_NP)
otu_tab_to_df(ps_NP)
}
\seealso{
\code{\link{meta_to_df}}
}
