% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{otutax_to_dataframe}
\alias{otutax_to_dataframe}
\title{OTU taxa to dataframe}
\usage{
otutax_to_dataframe(phyloseq_obj)
}
\arguments{
\item{phyloseq_obj}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
A function for manipulating a phyloseq object into a list containing the following tibbles:
Tax table, OTU table, Metadata table, Stressed/Control Samples, OTU names, Merged Tax/OTU table
}
