% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{parse_taxonomy_silva}
\alias{parse_taxonomy_silva}
\title{Parse SILVA Taxonomy}
\usage{
parse_taxonomy_silva(char.vec)
}
\arguments{
\item{char.vec}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
This function parses SILVA data.
}
