% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phyloseq_silva_2}
\alias{phyloseq_silva_2}
\title{Phyloseq Data with 2 Treatment Groups}
\format{An object of class \code{"phyloseq"}}
\source{
\href{https://nephele.niaid.nih.gov/user_guide_pipes/#amplicon_pipes}{Nephele-16S Qiime Pipeline}
}
\usage{
data(phyloseq_silva_2)
}
\description{
Data from the Nephele-16S-Qiime-Silva microbiome pipeline
in the form of a phyloseq object.
}
\references{
Zhang et al. (2019)
(\href{https://www.ncbi.nlm.nih.gov/pubmed/}{PubMed})
}
\keyword{datasets}
