% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{phyloseq_to_dataframe}
\alias{phyloseq_to_dataframe}
\title{FUNCTION_TITLE}
\usage{
phyloseq_to_dataframe(phyloseq_obj, treatment_groups)
}
\arguments{
\item{phyloseq_obj}{PARAM_DESCRIPTION}

\item{treatment_groups}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[phyloseq]{tax_table}},\code{\link[phyloseq]{otu_table}},\code{\link[phyloseq]{sample_data}}
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{join}}
\code{\link[stringr]{case}}
}
\concept{FAMILY_TITLE}
