% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{phyloseq_to_excel}
\alias{phyloseq_to_excel}
\title{Convert Phyloseq Object to Statistical Excel File}
\usage{
phyloseq_to_excel(phyloseq_obj, treatment_groups, rank, file_path)
}
\arguments{
\item{phyloseq_obj}{A phyloseq object.}

\item{treatment_groups}{A list of the treatment groups in the sample metadata.}

\item{rank}{The agglomerated rank of the phyloseq object}

\item{file_path}{Absolute path of the Excel file that will be created.}
}
\value{
An Excel file containing all of the phyloseq data with additional statistical data.

No objects are returned, but an Excel file is created.
}
\description{
This function takes a phyloseq object and converts it to a list of
dataframes followed by analyzing the treatment groups and adding statistical data to
the same list.  It then creates an excel file from this object.
}
\seealso{
\code{\link[dplyr]{select_all}}
\code{\link[openxlsx]{createWorkbook}},\code{\link[openxlsx]{addWorksheet}},\code{\link[openxlsx]{writeDataTable}},\code{\link[openxlsx]{saveWorkbook}}

Other Data Manipulators: \code{\link{phyloseq_to_dataframe}},
  \code{\link{phyloseq_to_stats_dataframe}},
  \code{\link{phyloseq_to_tibble}}
}
\concept{Data Manipulators}
