% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{phyloseq_to_stats_dataframe}
\alias{phyloseq_to_stats_dataframe}
\title{Convert Phyloseq Object to Statistical Data Frame}
\usage{
phyloseq_to_stats_dataframe(phyloseq_obj)
}
\arguments{
\item{phyloseq_obj}{A phyloseq object.}
}
\value{
A list of dataframes containing all of the phyloseq data with additional statistical data.
}
\description{
This function takes a phyloseq object and converts it to a list of
dataframes followed by analyzing the treatment groups and adding statistical data to
the same list.
}
\examples{
\dontrun{
> phy_obj <- get_phyloseq_object(...)
> stats_df <- phyloseq_to_stats_dataframe(phy_obj, treatment_groups = list("Stressed", "Control"))
> stats_df$stats
TODO: Add a stats dataframe here
}
}
\seealso{
Other Data Manipulators: \code{\link{metacoder_comp_func_1}},
  \code{\link{metacoder_workflow_1}},
  \code{\link{phyloseq_to_dataframe}},
  \code{\link{phyloseq_to_excel}},
  \code{\link{phyloseq_to_tibble}}, \code{\link{vlookup}}
}
\concept{Data Manipulators}
