% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{pick_new_outgroup}
\alias{pick_new_outgroup}
\title{Pick New Outgroup}
\usage{
pick_new_outgroup(unrooted_tree)
}
\arguments{
\item{unrooted_tree}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
The tree rooting functions inspired by recent comments in https://github.com/joey711/phyloseq/issues/597
}
