% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{preprocess_phyloseq}
\alias{preprocess_phyloseq}
\title{Preprocess Phyloseq}
\usage{
preprocess_phyloseq(phyloseq_object, process_list = NULL)
}
\arguments{
\item{phyloseq_object}{DESCRIPTION.}

\item{process_list}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
The preprocessing function
}
