% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_to_df.R
\name{ps_to_df}
\alias{ps_to_df}
\title{Combine phyloseq OTU-table (otu_table) and metadata (sample_data) into one dataframe.}
\usage{
ps_to_df(ps, sample_name = "sample_id")
}
\arguments{
\item{ps}{phyloseq-object}

\item{sample_name}{name of the column containing samples (default: 'sample_id')}
}
\value{
A dataframe with OTU-data and metadata contained in phyloseq.
}
\description{
Combine phyloseq OTU-table (otu_table) and metadata (sample_data) into one dataframe.
}
\examples{
data(ps_NP)
ps_to_df(ps_NP)
}
\seealso{
\code{\link{meta_to_df}} \code{\link{otu_tab_to_df}}
}
