% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_greengenes}
\alias{raw_greengenes}
\title{Stress Microbiome "raw_format" Data (Greengenes)}
\format{An object of class \code{"Taxmap"} with customized observation tables:
\describe{
\item{otu_abundance}{An abundance table with otu_id, taxon_id, and Sample_* columns.}
\item{otu_annotations}{A taxonomy annotations table with otu_id, taxon_id, and multiple rank columns.}
\item{sample_data}{A metadata table with SampleID, BarcodeSequence, LinkerPrimerSequence,
ForwardFastqFile, ReverseFastqFile, TreatmentGroup, SampleName, Description}
\item{phy_tree}{A phylogenetic tree.}
}}
\source{
\href{https://nephele.niaid.nih.gov/user_guide_pipes/#amplicon_pipes}{Nephele-16S Qiime Pipeline}
}
\usage{
data(raw_greengenes)
}
\description{
Data from the Nephele-16S-Qiime-Greengenes microbiome pipeline
in the form of a taxmap object.  This taxmap object has been
formatted into the MicrobiomeR "raw_format".
}
\references{
Zhang et al. (2019)
(\href{https://www.ncbi.nlm.nih.gov/pubmed/}{PubMed})
}
\keyword{datasets}
