% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{remove_ambiguous_taxa}
\alias{remove_ambiguous_taxa}
\title{Remove Ambiguous Taxa}
\usage{
remove_ambiguous_taxa(phyloseq_obj, ranks, ambiguous_names)
}
\arguments{
\item{phyloseq_obj}{DESCRIPTION.}

\item{ranks}{DESCRIPTION.}

\item{ambiguous_names}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
A function for removing named ambiguous taxonomy at specific rank(s)
}
