% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{root_by_longest_edge}
\alias{root_by_longest_edge}
\title{Root by Longest Edge}
\usage{
root_by_longest_edge(unrooted_tree)
}
\arguments{
\item{unrooted_tree}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
