% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{root_phyloseq_tree}
\alias{root_phyloseq_tree}
\title{Root phyloseq tree.}
\usage{
root_phyloseq_tree(phyloseq_object, tree_path, save_rooted_tree,
  recursive = TRUE)
}
\arguments{
\item{phyloseq_object}{A phyloseq object that contains a phylogenetic tree.}

\item{tree_path}{The path to the existing or desired phylogenetic tree file.}

\item{save_rooted_tree}{A logical that determines weather or not the rooted tree
is saved.}

\item{recursive}{If the directory doesn't exists create the parent directories
that don't exist as well, Default: TRUE}
}
\value{
Returns a phyloseq object with a rooted tree.
}
\description{
A function for rooting and saving a phylogenetic tree.
}
\details{
This function is a helper function to get a proper phyloseq object for
downstream analysis.  Some analysis requre a rooted tree.  The function saves
the rooted tree in the phyloseq object.  It can also save the tree as a file
if desired.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ape]{root}}, \code{\link[ape]{read.tree}}, \code{\link[ape]{write.tree}}

\code{\link[phyloseq:phy_tree-methods]{phy_tree}}

\code{\link[tools]{fileutils}}

Other Phylogenetic Tree Manipulators: \code{\link{pick_new_outgroup}},
  \code{\link{root_by_longest_edge}}
}
\concept{Phylogenetic Tree Manipulators}
