% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{save_barplot}
\alias{save_barplot}
\title{Save Barplot}
\usage{
save_barplot(plot, filename)
}
\arguments{
\item{plot}{The stacked barplot object.}

\item{filename}{The filename of the plot. (It will be saved as .tiff)}
}
\description{
Save a stacked barplot.
}
