% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{save_top_coefficients_barplot}
\alias{save_top_coefficients_barplot}
\title{Save Top Coefficients Barplot}
\usage{
save_top_coefficients_barplot(plot, filename, format = "tiff",
  start_path = "output", ...)
}
\arguments{
\item{plot}{The plot object.}

\item{filename}{The name of the file. (an extension should not be included)}
}
\description{
Save a top coefficients barplot.
}
\examples{
\dontrun{
if (interactive()) {
  library(MicrobiomeR)
  save_top_coefficients_barplot(plot = plot, filename = "stacked_bar_phylum")
}
}
}
