% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scico_palette}
\alias{scico_palette}
\title{Scico Palette Function}
\usage{
scico_palette(scico_palette = "batlow", scico_number = 800,
  scico_range = 0:scico_number)
}
\arguments{
\item{scico_palette}{The scico palette to use.  Default: 'batlow'}

\item{scico_number}{The number of colors to use in the scico palette.  Default: 800}

\item{scico_range}{The range in the color palette to use.  Default: c(0, scico_number)}
}
\value{
The output of this function is another function (grDevoces::colorRampPalette), which takes
a number to generate an interpolated color palette as a character vector.
}
\description{
A function that returns a color palette function based off of the scico package.
}
\details{
The purpose of this function is to provide an interpolated scico palette for using with the
get_color_palette function.
}
\examples{
\dontrun{
if(interactive()){
library(Microbiome)
basic_pal <- scico_palette(scico_palette="hawaii")(25)
better_pal <- get_color_palette(pal_func = scico_palette, color_no = 25)
 }
}
}
\seealso{
\code{\link[grDevices]{colorRamp}}
\code{\link[scico]{scico}}

Other Color Palettes: \code{\link{combination_palette}},
  \code{\link{get_color_palette}},
  \code{\link{viridis_magma_palette}},
  \code{\link{viridis_palette}}
}
\concept{Color Palettes}
