% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scico_palette_func}
\alias{scico_palette_func}
\title{Scico Palette Function}
\usage{
scico_palette_func(scico_palette = "batlow", scico_number = 800,
  scico_range = c(0, scico_number))
}
\arguments{
\item{scico_palette}{The scico palette to use.  Default: 'batlow'}

\item{scico_number}{The number of colors to use in the scico palette.  Default: 800}

\item{scico_range}{The range in the color palette to use.  Default: c(0, scico_number)}
}
\value{
The output of this function is another function (grDevoces::colorRampPalette), which takes
a number to generate an interpolated color palette as a character vector.
}
\description{
A function that returns a color palette function based off of the scico package.
}
\details{
The purpose of this function is to provide an interpolated scico palette for using with the
get_color_palette function.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[grDevices]{colorRamp}}
\code{\link[scico]{scico}}

Other Color Palettes: \code{\link{combo_palette_func}},
  \code{\link{get_color_palette}},
  \code{\link{viridis_palette_func}},
  \code{\link{virmag_palette_func}}
}
\concept{Color Palettes}
