% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skewness}
\alias{skewness}
\title{FUNCTION_TITLE}
\usage{
skewness(x, finite = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{finite}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
