% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{stacked_barplot}
\alias{stacked_barplot}
\title{Stacked Barplot}
\usage{
stacked_barplot(
  obj,
  tax_level = "Phylum",
  fill = "Phylum",
  xlabel = "Samples",
  faceted = FALSE,
  title = NULL,
  palette_values = NULL
)
}
\arguments{
\item{obj}{An object to be converted to a taxmap object with \code{\link[MicrobiomeR]{create_taxmap}}.}

\item{tax_level}{The taxonomic level, Default: 'Phylum'}

\item{fill}{The taxonomic level by which the bars are filled, Default: 'Phylum'}

\item{xlabel}{The label of the x axis, Default: 'Samples'}

\item{faceted}{A boolean to determine if the barplot should be faceted by TreatmentGroup}

\item{title}{The title or name of the plot.}

\item{palette_values}{A list of the colors to input to be mapped to the plot palette, Default: 'NULL'}
}
\value{
Returns a stacked barplot.
}
\description{
Create a stacked barplot to show relative abundance of taxa. \code{convert_proportions} converts the dataframe abundance values to percent 100 and returns a transformed dataframe.
\code{melt_metacoder} melts the metacoder or phyloseq tables into a dataframe and returns a melted dataframe. \code{stacked_barplots} creates a stacked barplots for multiple taxonomic levels and returns a list of stacked barplots.
}
\examples{
\dontrun{
if (interactive()) {
  # An example stacked bar plot
  library(MicrobiomeR)
  data <- analyzed_silva
  palette <- get_color_palette(color_no = 12)
  plot <- stacked_barplot(obj = data, palette_values = palette)
  plot
}
}
}
\seealso{
Other Visualizations: 
\code{\link{alpha_diversity_plot}()},
\code{\link{correlation_data}()},
\code{\link{correlation_plot}()},
\code{\link{correlation_plots}()},
\code{\link{heat_tree_parameters}()},
\code{\link{heat_tree_plots}()},
\code{\link{ordination_plot}()},
\code{\link{plot_limits}()},
\code{\link{save_alpha_diversity_plots}()},
\code{\link{save_correlation_plots}()},
\code{\link{save_heat_tree_plots}()},
\code{\link{save_ordination_plots}()},
\code{\link{save_stacked_barplots}()},
\code{\link{stacked_barplots}()},
\code{\link{top_coefficients_barplot}()}
}
\concept{Visualizations}
