% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{stacked_barplots}
\alias{stacked_barplots}
\title{Stacked Barplots}
\usage{
stacked_barplots(obj, tax_levels = c("Phylum", "Class", "Order"))
}
\arguments{
\item{obj}{An object to be converted to a taxmap object with \code{\link[MicrobiomeR]{create_taxmap}}.}

\item{tax_levels}{The taxonomic levels, Default: 'c("Phylum", "Class", "Order")'}
}
\description{
Stacked Barplots
}
\seealso{
Other Visualizations: 
\code{\link{alpha_diversity_plot}()},
\code{\link{correlation_data}()},
\code{\link{correlation_plot}()},
\code{\link{correlation_plots}()},
\code{\link{heat_tree_parameters}()},
\code{\link{heat_tree_plots}()},
\code{\link{ordination_plot}()},
\code{\link{plot_limits}()},
\code{\link{save_alpha_diversity_plots}()},
\code{\link{save_correlation_plots}()},
\code{\link{save_heat_tree_plots}()},
\code{\link{save_ordination_plots}()},
\code{\link{save_stacked_barplots}()},
\code{\link{stacked_barplot}()},
\code{\link{top_coefficients_barplot}()}
}
\concept{Visualizations}
