% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{top_coefficients_barplot}
\alias{top_coefficients_barplot}
\title{Top Coefficients Barplot}
\usage{
top_coefficients_barplot(top_coefficients, title = NULL)
}
\arguments{
\item{top_coefficients}{Coefficients for the top taxa separating the groups.}

\item{title}{The title of the plot, Default = NULL}
}
\value{
Returns a barplot of the top coefficients
}
\description{
A barplot to visualize which taxa have had the greatest impact on the overall gut composition.
}
\examples{
\dontrun{
if (interactive()) {
  library(MicrobiomeR)
  data <- analyzed_silva
  p <- permanova(obj = data)
  top_coefficients_barplot(p$top_coefficients)
}
}
}
\seealso{
View \code{\link{permanova}} to understand how to generate \code{top_coefficents}

Other Visualizations: 
\code{\link{alpha_diversity_plot}()},
\code{\link{correlation_data}()},
\code{\link{correlation_plot}()},
\code{\link{correlation_plots}()},
\code{\link{heat_tree_parameters}()},
\code{\link{heat_tree_plots}()},
\code{\link{ordination_plot}()},
\code{\link{plot_limits}()},
\code{\link{save_alpha_diversity_plots}()},
\code{\link{save_correlation_plots}()},
\code{\link{save_heat_tree_plots}()},
\code{\link{save_ordination_plots}()},
\code{\link{save_stacked_barplots}()},
\code{\link{stacked_barplot}()},
\code{\link{stacked_barplots}()}
}
\concept{Visualizations}
