% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{transform_metacoder}
\alias{transform_metacoder}
\title{Transform Metacoder Dataframe}
\usage{
transform_metacoder(melted_df, tax_level)
}
\arguments{
\item{melted_df}{A "melted" dataframe from the metacoder object's data.}

\item{tax_level}{The taxonomic level.}
}
\value{
Returns a transformed dataframe.
}
\description{
Transform the dataframe abundance values to percent 100.
}
