% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{transform_metacoder_obj}
\alias{transform_metacoder_obj}
\title{Transform Metacoder Object}
\usage{
transform_metacoder_obj(melted_metacoder_df, tax_level)
}
\arguments{
\item{melted_metacoder_df}{DESCRIPTION.}

\item{tax_level}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Transform the dataframe abundance values.
}
