% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformer}
\alias{transformer}
\title{Transforming Tidy Data}
\usage{
transformer(.data, func, by = "column", ids = NULL,
  header_name = NULL, preserved_categories = TRUE,
  separated_categories = NULL, ...)
}
\arguments{
\item{.data}{A matrix/data_frame/tibble for transforming.}

\item{func}{A function, which can be anonymous, that will be used to transform the data.  (e.g. proportions
x/sum(x)).}

\item{by}{This denotes how the data should be transformed (column/row). Default: 'column'}

\item{ids}{The column to transpose by. Default: The first column.}

\item{header_name}{A name for the numeric data that will be transposed.}

\item{preserved_categories}{A logical denoting weather categorical data should be conserved.  A
value of FALSE will cause all categorical data except the \emph{ids} to be dropped.  A value of
TRUE will cause the categorical data to preserved by \emph{tidyr::unite}ing these columns.  Default: TRUE}

\item{separated_categories}{A vector containing ordered column names to use in a previously transposed
and categorically preserved table.  Retransposing with this set should yield an exact replicate of
the original data.  Default: NULL}

\item{...}{Additional arguments passed on to \emph{func}}
}
\value{
A tibble that has been transformed.
}
\description{
This function transforms a table by rows or by columns.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{transposer}}
\code{\link[dplyr]{select_all}},\code{\link[dplyr]{select}}
\code{\link[purrr]{modify}}
}
\concept{Data Manipulator}
