% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{treatment_matrix}
\alias{treatment_matrix}
\title{Get Treatment Matrix}
\usage{
treatment_matrix(obj)
}
\arguments{
\item{obj}{An object to be converted to a taxmap object with \code{\link[MicrobiomeR]{create_taxmap}}.}
}
\value{
A matrix with each column representing a comparison to be made.
}
\description{
A function that returns a matrix with used for comparing treatment data.
}
\details{
Use this when you want to do pairwise comparisons.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other Formatting: 
\code{\link{MicrobiomeR_Formats}},
\code{\link{as_MicrobiomeR_format}()},
\code{\link{as_analyzed_format}()},
\code{\link{as_basic_format}()},
\code{\link{as_custom_format}()},
\code{\link{as_phyloseq_format}()},
\code{\link{as_raw_format}()},
\code{\link{melt_taxmap}()}
}
\concept{Formatting}
