% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{validate_MicrobiomeR_format}
\alias{validate_MicrobiomeR_format}
\title{Validate MicrobiomeR Format}
\usage{
validate_MicrobiomeR_format(obj, validated = FALSE, valid_formats,
  force_format = FALSE, min_or_max = base::max, ...)
}
\arguments{
\item{obj}{A Taxmap object.}

\item{validated}{This parameter provides a way to override validation steps.  Use carefully.  Default: FALSE}

\item{valid_formats}{A vector of formats that are used for validation.}

\item{force_format}{A logical denoting if the selected format is to be forced.  Default: FALSE}

\item{min_or_max}{A base function (min or max) that determines which format is to be forced.
This is particularly useful if you provide multiple \emph{valid_formats}.  Min will choose the lowest
level format, while max will choose the highest level format.  Default: base::max}

\item{...}{An optional list of parameters to use in \code{\link[MicrobiomeR]{as_MicrobiomeR_format}}.}
}
\value{
If the object is validated, a Taxmap object.
}
\description{
This funciton validates that Taxmap objects are in a valid format MicrobiomeR format.
}
\details{
This function can provide a way to check if a taxmap object has undergone a
\code{\link[MicrobiomeR:MicrobiomeR_Workflow]{MicrobiomeR Style Workflow}}.
}
\seealso{
\code{\link[MicrobiomeR]{which_format}},  \code{\link[MicrobiomeR]{as_MicrobiomeR_format}}

Other Validation: \code{\link{is_analyzed_format}},
  \code{\link{is_basic_format}},
  \code{\link{is_phyloseq_format}},
  \code{\link{is_raw_format}},
  \code{\link{order_metacoder_data}},
  \code{\link{which_format}}
}
\concept{Validation}
