% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{virdis_palette_func}
\alias{virdis_palette_func}
\title{Viridis Palette Function}
\usage{
virdis_palette_func(viridis_number = 800, viridis_range = c(300, 800),
  magma_number = 500, magma_range = c(0, 500))
}
\arguments{
\item{viridis_number}{The total number of colors used to generate the viridis palette. Default: 800}

\item{viridis_range}{Tne range of colors in the viridis palette to use. Default: c(300, 800)}

\item{magma_number}{The total number of colors used to generate the magma palette. Default: 500}

\item{magma_range}{The range of colors in the magma palette to use. Default: c(0, 500)}
}
\value{
The output of this function is another function, which takes a number to generate a color
palette as a character vector.
}
\description{
A function that returns a color palette function based off of the veridis package.
}
\details{
The primary purpose of this function is to return a palette-function for generating virdis style
color palettes.  By taking the viridis::viridis() and the viridis::magma() colors, and manipulating
them, this function can help create a unique set of colors that you can distinguish on a busy plot.
The hopes of this function is to help improve plots that use more than 20 colors.  Use the provided
example to view the color palette.
}
\examples{
\dontrun{
if(interactive()){
 # Use the default values
 > pal_func <- viridis_palette_func()

 # Get a palette with 20 colors
 > pal <- pal_func(20)

 # Make a pie plot of the colros.
 > pie(rep(1, length(pal)), col=pal)
 }
}
}
\seealso{
\code{\link[viridis]{reexports}}
\code{\link[MicrobiomeR]{get_color_palette}}

Other Color Palettes: \code{\link{get_color_palette}}
}
\concept{Color Palettes}
