% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{viridis_palette}
\alias{viridis_palette}
\title{Viridis Palette Function}
\usage{
viridis_palette(viridis_palette = "viridis", viridis_number = 800,
  viridis_range = 0:viridis_number)
}
\arguments{
\item{viridis_palette}{The viridis palette to use.  Default: 'batlow'}

\item{viridis_number}{The number of colors to use in the viridis palette.  Default: 800}

\item{viridis_range}{The range in the color palette to use.  Default: c(0, viridis_number)}
}
\value{
The output of this function is another function (grDevoces::colorRampPalette), which takes
a number to generate an interpolated color palette as a character vector.
}
\description{
A function that returns a color palette function based off of the viridis package.
}
\details{
The purpose of this function is to provide an interpolated viridis palette for using with the
get_color_palette function.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[grDevices]{colorRamp}}
\code{\link[viridis:reexports]{viridis}}

Other Color Palettes: \code{\link{combination_palette}},
  \code{\link{get_color_palette}},
  \code{\link{scico_palette}},
  \code{\link{viridis_magma_palette}}
}
\concept{Color Palettes}
