% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vlookup}
\alias{vlookup}
\title{Mock Excel "VlookUp" Function}
\usage{
vlookup(lookup_vector, df, match_var, return_var)
}
\arguments{
\item{lookup_vector}{A vector of items to look up.}

\item{df}{The dataframe to search.}

\item{match_var}{The column name to search in the dataframe.}

\item{return_var}{The column of data to return when matched.}
}
\value{
A vector that contains the items of interest.
}
\description{
A function that mimics excels vlookup, but for R's dataframe.
}
\details{
A function that works like the VLOOKUP function in excel.  This function was
borrowed from \url{https://www.r-bloggers.com/an-r-vlookup-not-so-silly-idea/}.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other Data Manipulators: \code{\link{convert_proportions}},
  \code{\link{metacoder_comp_func_1}},
  \code{\link{transformer}}, \code{\link{transposer}}
}
\concept{Data Manipulators}
