% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-diversity.R
\name{alpha_diversity_plot}
\alias{alpha_diversity_plot}
\title{Alpha Diversity Plot}
\usage{
alpha_diversity_plot(obj, measure = "shannon",
  select_otu_table = "otu_proportions", save = FALSE)
}
\arguments{
\item{obj}{An object to be converted to a metacoder object with \code{\link[MicrobiomeR]{object_handler}}.}

\item{measure}{Select an alpha diversity measure such as Shannon, Gini Simpson, and Inverse Simpson, Default: 'shannon'}

\item{select_otu_table}{Choose an otu table to analyze, Default: 'otu_proportions'}

\item{save}{Save the plot, Default: FALSE}
}
\value{
Returns an alpha diversity plot.
}
\description{
Plot the alpha diversity using a violin plot.
}
\details{
Alpha diversity helps to determine the species richness (the number of different species in a sample) or evenness (similar abundance level).
}
\examples{
\dontrun{
if (interactive()) {
  library(MicrobiomeR)
  data <- analyzed_silva
  plot <- alpha_diversity_plot(obj = data, measure = "shannon", select_otu_table = "otu_proportions", save = FALSE)
  plot
}
}
}
\seealso{
Other Visualizations: \code{\link{correlation_plot}},
  \code{\link{get_correlation_plots}},
  \code{\link{get_heat_tree_parameters}},
  \code{\link{get_heat_tree_plots}},
  \code{\link{get_plot_limits}},
  \code{\link{save_barplot}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{stacked_barplot}}
}
\concept{Visualizations}
