% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{as_analyzed_format}
\alias{as_analyzed_format}
\title{As Analyzed MicrobiomeR Format}
\usage{
as_analyzed_format(obj, cols = NULL, groups = NULL,
  combinations = NULL, out_names = NULL,
  comp_func = metacoder_comp_func_1)
}
\arguments{
\item{obj}{An object to be converted to a metacoder object with \code{\link[MicrobiomeR]{object_handler}}.}

\item{cols}{Column names used for \code{\link[metacoder]{calc_taxon_abund}}.  Default: NULL}

\item{groups}{Group names used for \code{\link[metacoder]{compare_groups}}.  Default: NULL}

\item{combinations}{Combinations of treatments used in \code{\link[metacoder]{compare_groups}}.  Default: NULL}

\item{out_names}{Column names of the output used for \code{\link[metacoder]{calc_obs_props}}.  Default: NULL}

\item{comp_func}{A Comparison based function used in \code{\link[metacoder]{compare_groups}}.  Default: NULL}
}
\value{
A Taxmap/metacoder object in the "analyzed_format".
}
\description{
Converts a metacoder object to the analyzed_format.
}
\details{
See the \link{MicrobiomeR_Formats} documentation.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{is_phyloseq_format}}, \code{\link[MicrobiomeR]{is_raw_format}},  \code{\link[MicrobiomeR]{is_basic_format}},  \code{\link[MicrobiomeR]{is_analyzed_format}},  \code{\link[MicrobiomeR]{as_raw_format}},  \code{\link[MicrobiomeR]{as_basic_format}},  \code{\link[MicrobiomeR]{order_metacoder_data}}
\code{\link[metacoder]{compare_groups}}
\code{\link[taxa]{taxonomy_table}}

Other Formatting: \code{\link{as_MicrobiomeR_format}},
  \code{\link{as_basic_format}},
  \code{\link{as_phyloseq_format}},
  \code{\link{as_raw_format}},
  \code{\link{format_metacoder_object}}
}
\concept{Formatting}
