% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{as_basic_format}
\alias{as_basic_format}
\title{As Basic MicrobiomeR Format}
\usage{
as_basic_format(obj, cols = NULL, out_names = NULL)
}
\arguments{
\item{obj}{A Taxmap/metacoder object.}

\item{cols}{Column names used for \code{\link[metacoder]{calc_taxon_abund}}.  Default: NULL}

\item{out_names}{Column names of the output used for \code{\link[metacoder]{calc_obs_props}}.  Default: NULL}
}
\value{
A Taxmap/metacoder object in the "basic_format".
}
\description{
Converts a metacoder object to the basic_format.
}
\details{
See the \link{MicrobiomeR_Formats} documentation.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{is_phyloseq_format}},  \code{\link[MicrobiomeR]{is_raw_format}},  \code{\link[MicrobiomeR]{is_basic_format}}, \code{\link[MicrobiomeR]{as_raw_format}},  \code{\link[MicrobiomeR]{order_metacoder_data}}
\code{\link[metacoder]{calc_taxon_abund}}, \code{\link[metacoder]{calc_obs_props}}

Other Formatting: \code{\link{as_MicrobiomeR_format}},
  \code{\link{as_analyzed_format}},
  \code{\link{as_raw_format}}
}
\concept{Formatting}
