% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{as_raw_format}
\alias{as_raw_format}
\title{As Raw MicrobiomeR Format}
\usage{
as_raw_format(obj)
}
\arguments{
\item{obj}{An object to be converted to a metacoder object with \code{\link[MicrobiomeR]{object_handler}}.}
}
\value{
A Taxmap/metacoder object in the "raw_format".
}
\description{
Converts a metacoder object to the raw_format.
}
\details{
See the \link{MicrobiomeR_Formats} documentation.
}
\seealso{
\code{\link[MicrobiomeR]{is_phyloseq_format}}, \code{\link[MicrobiomeR]{is_raw_format}}, \code{\link[MicrobiomeR]{order_metacoder_data}}

Other Formatting: \code{\link{as_MicrobiomeR_format}},
  \code{\link{as_analyzed_format}},
  \code{\link{as_basic_format}},
  \code{\link{as_phyloseq_format}},
  \code{\link{format_metacoder_object}}
}
\concept{Formatting}
