% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{combination_palette}
\alias{combination_palette}
\title{Combine Color Palettes}
\usage{
combination_palette(...)
}
\arguments{
\item{...}{You can use any name for your arguments, but the values must be a named list.
The list can only have 4 named members:
\describe{
\item{palette}{This is a palette function that returns a vector of colors.}
\item{args}{This is another named list used for the palette function parameters.}
\item{range}{This is a range \emph{(1:10)} used to subset the color palette vector.}
\item{rev}{This is a logical \emph{(TRUE/FALSE)} used to reverse the color palette.}
}
You can add as many parameters you want in order to combine as many color palettes
as you want.}
}
\value{
The output of this function is another function (grDevoces::colorRampPalette), which takes
a number to generate an interpolated color palette as a character vector.
}
\description{
This function uses dynamic arguments (...) in order to combine multiple
color palettes together.
}
\details{
This function allows you to combine a varying number of color palettes and gives you
the ability to subset and reverse the palettes that are supplied.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[grDevices]{colorRamp}}

Other Color Palettes: \code{\link{get_color_palette}},
  \code{\link{scico_palette}},
  \code{\link{viridis_magma_palette}},
  \code{\link{viridis_palette}}
}
\concept{Color Palettes}
