% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-plot.R
\name{correlation_data}
\alias{correlation_data}
\title{Get Correlation Plot Data}
\usage{
correlation_data(obj, primary_rank, secondary_rank = TRUE, wp_value = 0.05)
}
\arguments{
\item{obj}{An object to be converted to a Taxmap object with \code{\link[MicrobiomeR]{create_taxmap}}.}

\item{primary_rank}{A primary rank used to label the points.}

\item{secondary_rank}{The secondary rank used to color the points, Default: TRUE}

\item{wp_value}{The wilcoxon p-value cutoff/threshold, Default: 0.05}
}
\value{
Data for the correlation plots.
}
\description{
Get the correlation plot data comparing all of the treatment groups.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{agglomerate_taxmap}},  \code{\link[MicrobiomeR]{vlookup}}

\code{\link[dplyr]{tidyeval}},  \code{\link[dplyr]{mutate}},  \code{\link[dplyr]{filter}},  \code{\link[dplyr]{arrange}}

Other Visualizations: 
\code{\link{alpha_diversity_plot}()},
\code{\link{correlation_plot}()},
\code{\link{correlation_plots}()},
\code{\link{heat_tree_parameters}()},
\code{\link{heat_tree_plots}()},
\code{\link{ordination_plot}()},
\code{\link{plot_limits}()},
\code{\link{save_alpha_diversity_plots}()},
\code{\link{save_correlation_plots}()},
\code{\link{save_heat_tree_plots}()},
\code{\link{save_ordination_plots}()},
\code{\link{save_stacked_barplots}()},
\code{\link{stacked_barplot}()},
\code{\link{stacked_barplots}()},
\code{\link{top_coefficients_barplot}()}
}
\concept{Visualizations}
