% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plot.R
\name{correlation_plot}
\alias{correlation_plot}
\title{Correlation Plots}
\usage{
correlation_plot(obj, primary_rank, secondary_rank = TRUE,
  wp_value = 0.05)
}
\arguments{
\item{obj}{An object to be converted to a metacoder object with \code{\link[MicrobiomeR]{object_handler}}.}

\item{primary_rank}{The primary rank used to label the points.}

\item{secondary_rank}{The secondary rank used to color the points.  Default: TRUE}

\item{wp_value}{The Wilcoxian P-Value used to represent significant points.  Default: 0.05}
}
\value{
A 1:1 correlation plot built with ggplot2.
}
\description{
Create a correlation plot from a metacoder/taxmap object.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{object_handler}},  \code{\link[MicrobiomeR]{validate_MicrobiomeR_format}},  \code{\link[MicrobiomeR]{agglomerate_metacoder}},  \code{\link[MicrobiomeR]{vlookup}},  \code{\link[MicrobiomeR]{get_plot_limits}},  \code{\link[MicrobiomeR]{get_color_palette}}
\code{\link[dplyr]{tidyeval}},  \code{\link[dplyr]{mutate}},  \code{\link[dplyr]{filter}},  \code{\link[dplyr]{arrange}}
\code{\link[ggplot2]{ggplot}},  \code{\link[ggplot2]{aes}},  \code{\link[ggplot2]{geom_polygon}},  \code{\link[ggplot2]{geom_point}},  \code{\link[ggplot2]{labs}},  \code{\link[ggplot2]{scale_continuous}},  \code{\link[ggplot2]{scale_manual}},  \code{\link[ggplot2]{guide_legend}},  \code{\link[ggplot2]{geom_abline}}
\code{\link[ggrepel]{geom_label_repel}}
\code{\link[forcats]{fct_reorder}}

Other Visualizations: \code{\link{get_plot_limits}}
}
\concept{Visualizations}
