% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-plot.R
\name{correlation_plots}
\alias{correlation_plots}
\title{Get Multiple Correlation Plots}
\usage{
correlation_plots(obj, primary_ranks, secondary_ranks = TRUE, ...)
}
\arguments{
\item{obj}{An object to be converted to a Taxmap object with \code{\link[MicrobiomeR]{create_taxmap}}.}

\item{primary_ranks}{A vector of primary ranks used to label the points.}

\item{secondary_ranks}{The secondary rank used to color the points.  Can be an integer specifying
the number of supertaxon ranks above the primary rank or the name of a supertaxon rank.  Default: TRUE}

\item{...}{An optional list of parameters to use in the correlation_plot function.}
}
\value{
A list object containing correlation plots.  A pairwise comparison returns a nested list.
}
\description{
This function allows the user to create a list of multiple correlation plots.
}
\details{
This function makes it easier to generate multiple correlation plots at once.
}
\examples{
\dontrun{
if(interactive()){
# This example uses data that are no longer available in the MicrobiomeR package,
# however, they can be easily generated with \\code{\\link{MicrobiomeR}{as_analyzed_format}}.
library(MicrobiomeR)
analyzed_silva <- as_MicrobiomeR_format(MicrobiomeR::raw_silva_2, "analyzed_format")
corr_plots <- correlation_plots(analyzed_silva, primary_ranks = c("Phylum", "Class", "Order"),
                      secondary_ranks = c("Phylum", "Class", "Order", "Family", "Genus"))
# Show a plot
corr_plots$Class$Phylum
 }
}
}
\seealso{
Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_data}},
  \code{\link{correlation_plot}},
  \code{\link{heat_tree_parameters}},
  \code{\link{heat_tree_plots}}, \code{\link{melt_taxmap}},
  \code{\link{ordination_plot}}, \code{\link{plot_limits}},
  \code{\link{save_alpha_diversity_plots}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{save_ordination_plots}},
  \code{\link{save_stacked_barplots}},
  \code{\link{top_coefficients_barplot}}
}
\concept{Visualizations}
