% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-diversity.R
\name{get_alpha_diversity}
\alias{get_alpha_diversity}
\title{Get Alpha Diversity}
\usage{
get_alpha_diversity(phyloseq_object)
}
\arguments{
\item{phyloseq_object}{A preprocessed phyloseq object.}
}
\value{
Returns a list object of alpha diversity measurements.
}
\description{
This function generates various alpha diversity measures include Shannon, Fisher, Coverage, Gini Simpson, and Inverse Simpson.
}
