% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Get a Color Palette}
\usage{
get_color_palette(
  pal_func = viridis_magma_palette,
  color_no = 20,
  display = TRUE,
  rev = FALSE,
  ...
)
}
\arguments{
\item{pal_func}{A function that \emph{returns the output function} from \code{\link[grDevices]{colorRamp}} or
a \emph{the output of the output function} from \code{\link[grDevices]{colorRamp}}.  Default: viridis_magma_palette}

\item{color_no}{The number of colors in the palette.  Default: 20}

\item{display}{A logical for displaying a pie chart of the palette. Default: TRUE}

\item{rev}{A logical that will reverse the order of the color palette.}

\item{...}{Parameters for the \emph{pal_func}.}
}
\value{
Returns a color palette in the form of a character vector.
}
\description{
Get a color palette with a specific number of colors.
}
\details{
This function is meant to be a plugin style function for users to created palettes.
}
\examples{
\dontrun{
if(interactive()){
# This example uses data that are no longer available in the MicrobiomeR package,
# however, they can be easily generated with \code{\link{MicrobiomeR}{as_analyzed_format}}.
 library(MicrobiomeR)
analyzed_silva <- as_MicrobiomeR_format(MicrobiomeR::raw_silva_2, "analyzed_format")
 data <- analyzed_silva$data$stats_tax_data$Phylum
 data_len <- length(unique(data))
 custom_pal <- get_color_palette(color_no=data_len)
 print(custom_pal)
 }
}
}
\seealso{
Other Color Palettes: 
\code{\link{combination_palette}()},
\code{\link{scico_palette}()},
\code{\link{viridis_magma_palette}()},
\code{\link{viridis_palette}()}
}
\concept{Color Palettes}
