% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-plot.R
\name{get_correlation_data}
\alias{get_correlation_data}
\title{Get Correlation Plot Data}
\usage{
get_correlation_data(obj, primary_rank, secondary_rank = TRUE,
  wp_value = 0.05)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{primary_rank}{PARAM_DESCRIPTION}

\item{secondary_rank}{PARAM_DESCRIPTION, Default: TRUE}

\item{wp_value}{PARAM_DESCRIPTION, Default: 0.05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get the correlation plot data comparing all of the treatment groups.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{agglomerate_metacoder}},  \code{\link[MicrobiomeR]{vlookup}}

\code{\link[dplyr]{tidyeval}},  \code{\link[dplyr]{mutate}},  \code{\link[dplyr]{filter}},  \code{\link[dplyr]{arrange}}

\code{\link[taxa]{filter_obs}}

Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_plot}},
  \code{\link{get_correlation_plots}},
  \code{\link{get_heat_tree_parameters}},
  \code{\link{get_heat_tree_plots}},
  \code{\link{get_plot_limits}},
  \code{\link{melt_metacoder}},
  \code{\link{ordination_plot}},
  \code{\link{save_barplot}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{save_ordination_plots}},
  \code{\link{top_coefficients_barplot}}
}
\concept{Visualizations}
