% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-plot.R
\name{get_correlation_plots}
\alias{get_correlation_plots}
\title{Get Multiple Correlation Plots}
\usage{
get_correlation_plots(obj, primary_ranks, secondary_ranks = TRUE,
  pairwise = FALSE, ...)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{pairwise}{This does a pairwise comparison of the primary and secondary ranks.  Default: FALSE}

\item{...}{An optional list of parameters to use in the correlation_plot function.}

\item{primary_rank}{A vector of primary ranks used to label the points.}

\item{secondary_rank}{The secondary rank used to color the points.  Can be an integer specifying
the number of supertaxon ranks above the primary rank or the name of a supertaxon rank.  Default: TRUE}
}
\value{
A list object containing correlation plots.  A pairwise comparison returns a nested list.
}
\description{
This function allows the user to create a list of multiple correlation plots.
}
\details{
This function makes it easier to generate multiple correlation plots at once.
}
\examples{
\dontrun{
if(interactive()){
library(MicrobiomeR)
corr_plots <- get_correlation_plots(MicrobiomeR::analyzed_silva, primary_ranks = c("Phylum", "Class", "Order"),
                      secondary_ranks = c("Phylum", "Class", "Order", "Family", "Genus"), pairwise = TRUE)
# Show a plot
corr_plots$Class$Phylum
 }
}
}
\seealso{
Other Visualizations: \code{\link{correlation_plot}},
  \code{\link{get_heat_tree_parameters}},
  \code{\link{get_heat_tree_plots}},
  \code{\link{get_plot_limits}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}}
}
\concept{Visualizations}
