% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plot.R
\name{get_correlation_plots}
\alias{get_correlation_plots}
\title{Get Multiple Correlation Plots}
\usage{
get_correlation_plots(obj, primary_ranks, secondary_ranks = TRUE,
  pairwise = FALSE, ...)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{pairwise}{This does a pairwise comparison of the primary and secondary ranks.  Default: FALSE}

\item{...}{An optional list of parameters to use in the correlation_plot function.}

\item{primary_rank}{A vector of primary ranks used to label the points.}

\item{secondary_rank}{The secondary rank used to color the points.  Can be an integer specifying
the number of supertaxon ranks above the primary rank or the name of a supertaxon rank.  Default: TRUE}
}
\value{
A list object containing correlation plots.  A pairwise comparison returns a nested list.
}
\description{
This function allows the user to create a list of multiple correlation plots.
}
\details{
This function makes it easier to generate multiple correlation plots at once.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other Visualizations: \code{\link{correlation_plot}},
  \code{\link{get_plot_limits}},
  \code{\link{save_correlation_plots}}
}
\concept{Visualizations}
