% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{get_distance_methods}
\alias{get_distance_methods}
\title{@title Get Distance Methods}
\usage{
get_distance_methods(tree_methods = TRUE)
}
\description{
@description The function for manipulating phyloseqs native distanceMethodList
}
\details{
@param tree_methods DESCRIPTION.

@return RETURN_DESCRIPTION
@export
}
