% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_output_dir}
\alias{get_output_dir}
\title{Create an Output Directory}
\usage{
get_output_dir(start_path = NULL, experiment = NULL,
  plot_type = NULL, end_path = NULL, root_path = NULL,
  custom_path = NULL, overwrite = FALSE, mkdir = TRUE)
}
\arguments{
\item{start_path}{With the start_path set, the full path generated by this function will be
\emph{root_path}/\emph{start_path}/\emph{?experiment?}/\emph{?plot_type?}/\emph{end_path}.  Default: NULL}

\item{experiment}{(optional) This will add a second level to the start_path file system.  Default: NULL}

\item{plot_type}{(optional) This will add a third level to the start_path file system.  With \strong{ONLY}
the plot_type set the full path generated by this function will be \emph{root_path}/\emph{plot_type}.  Default: NULL}

\item{end_path}{(optional) This will add the final directory to the start_path file system Default: A formatted date-time string.}

\item{root_path}{The root of the new file system (start_path or not).  Default: The working directory.}

\item{custom_path}{This is an absolute path that overrides everything else.  Default: NULL}

\item{overwrite}{A logical denoting that overwriting is acceptable.  Default: FALSE}

\item{mkdir}{A logical denoting weather or not the directory should be created or not.  Default: TRUE}
}
\value{
Creates a directory for output and returns the path as a string.
}
\description{
A function for generating a consistent file system for data files and visualizations.  If none
of the parameters are set the full path generated by this function will be \emph{root_path}/output
}
\details{
This function is increadibly useful on it's own but also for various other plotting/saving functions within the package.
It helps keep data organized using a standard workflow.
}
\examples{
\dontrun{
if(interactive()){
 # Get the path to an output directory without creating
 library(MicrobiomeR)
 output_dir <- get_output_dir(start_path="output", experiment="microbiome-proj", mkdir=FALSE)
 print(output_dir)

 # Create a folder for your plot types
 output_dir <- get_output_dir(plot_type="scatter-plots")
 print(output_dir)

 }
}
}
\seealso{
\code{\link[glue]{glue}}

\code{\link[stringr]{case}}
}
\concept{Project Management}
