% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plot.R
\name{get_plot_limits}
\alias{get_plot_limits}
\title{Get Plot Limits}
\usage{
get_plot_limits(x, y)
}
\arguments{
\item{x}{Data used to plot the x axis.}

\item{y}{Data used to plot the y axis.}
}
\value{
A vector that supplies the overarching x and y limits.
}
\description{
Get the limits of the plot using data values.
}
\seealso{
Other Visualizations: \code{\link{correlation_plot}}
}
\concept{Visualizations}
