% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{is_raw_format}
\alias{is_raw_format}
\title{Is Raw MicrobiomeR Format}
\usage{
is_raw_format(obj)
}
\arguments{
\item{obj}{A Taxmap/metacoder object.}
}
\value{
A logical (TRUE/FALSE).
}
\description{
This function returns a logical based on weather or not the object is in the raw_format.
}
\details{
The "raw_format" is Level 1. in the \link{MicrobiomeR_Formats} hierarchy.
}
\examples{
\dontrun{
if(interactive()){
library(MicrobiomeR)
if(is_raw_format(MicrobiomeR::raw_silva_2)){
    print("It's raw!")
} else {
    fmt <- which_format(MicrobiomeR::raw_silva_2)
    print("It's not analyzed!")
    print(sprintf("It's \%s!", fmt))
}
 }
}
}
\seealso{
Other Validation: \code{\link{is_analyzed_format}},
  \code{\link{is_basic_format}},
  \code{\link{is_phyloseq_format}},
  \code{\link{object_handler}},
  \code{\link{order_metacoder_data}},
  \code{\link{validate_MicrobiomeR_format}},
  \code{\link{which_format}}
}
\concept{Validation}
