% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{metacoder_comp_func_1}
\alias{metacoder_comp_func_1}
\title{Metacoder Comparison Function #1}
\usage{
metacoder_comp_func_1(abund_1, abund_2)
}
\arguments{
\item{abund_1}{A character vector of abundances.}

\item{abund_2}{A character vector of abundances.}
}
\value{
A list of statistical results used to compare groups.
}
\description{
A comparison function for metacoder::compare_groups "func" parameter.
}
\details{
This function is used by metacoder::compare_groups in order to compare
every combination of treatment groups.
}
\seealso{
\code{\link[diptest]{dip.test}}

\code{\link[modes]{bimodality_coefficient}}

Other Data Manipulators: \code{\link{transformer}},
  \code{\link{transposer}}, \code{\link{vlookup}}
}
\concept{Data Manipulators}
