% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomeR-package.R
\docType{package}
\name{MicrobiomeR}
\alias{MicrobiomeR}
\alias{MicrobiomeR-package}
\title{MicrobiomeR}
\description{
A microbiome workflow package using phyloseq for data import, metacoder for data analysis,
and ggplot2 for data visualization.
}
\details{
The goal of this package is to enhance the use of the \code{\link[metacoder:metacoder]{metacoder}} package by
providing tools that mimic the \code{\link[phyloseq:phyloseq-package]{phyloseq}} package.  The only difference being that
the main data object is a \code{\link[taxa:taxmap]{taxmap object}} instead of a \code{\link[phyloseq:phyloseq-class]{phyloseq object}}.
}
\section{Other Documentation}{

Below are some links to helpful R package used, while developing this package.

\describe{
\item{metacoder}{\url{https://grunwaldlab.github.io/metacoder_documentation/}}
\item{taxa}{\url{https://github.com/ropensci/taxa}}
\item{phyloseq}{\url{http://joey711.github.io/phyloseq/}}
\item{microbiome}{\url{http://microbiome.github.io/microbiome/}}
\item{MicrobiomeR}{\url{https://github.com/vallenderlab/MicrobiomeR}
}}
The \link{MicrobiomeR_Workflow} and \link{MicrobiomeR_Formats} help sections are also useful for understanding the
package.
}

\author{
Rob Gilmore and Shaurita Hutchins
}
