% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{object_handler}
\alias{object_handler}
\title{Object handler}
\usage{
object_handler(obj)
}
\arguments{
\item{obj}{An object that contains the data being analyzed.  Can be one of the following:
\describe{
\item{Phyloseq Object}{An object generated from the phyloseq package.}
\item{Taxmap Object}{An object generated from the metacoder or taxa package.}
\item{RData file}{An RData file generated from the base::save function.  Can have an extension of .RData or .rda.}
}}
}
\value{
The output generated is a \strong{taxa::taxmap} object.
}
\description{
A function that handles the conversion of objects to metacoder (\strong{taxa::taxmap}) objects.
}
\details{
This function is used to convert data to metacoder/taxmap objects for \emph{microbiome} analysis.
This function is used at the beginning of every other function to support multiple types of input
for the obj parameter in those functions.
}
\examples{
\dontrun{
if(interactive()){
 library(MicrobiomeR)
 phy_obj <- MicrobiomeR::phyloseq_silva
 mc_obj <- object_handler(phy_obj)
 }
}
}
\seealso{
\code{\link[metacoder]{parse_phyloseq}}

\code{\link[tools]{fileutils}}

Other Validation: \code{\link{is_analyzed_format}},
  \code{\link{is_basic_format}},
  \code{\link{is_phyloseq_format}},
  \code{\link{is_raw_format}},
  \code{\link{order_metacoder_data}},
  \code{\link{validate_MicrobiomeR_format}},
  \code{\link{which_format}}
}
\concept{Validation}
