% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{ordination_plot}
\alias{ordination_plot}
\title{Ordination Plot}
\usage{
ordination_plot(obj, method = "PCoA", distance = "wunifrac",
  title = NULL, save = FALSE, only_data = FALSE)
}
\arguments{
\item{obj}{An object to be converted to a metacoder object with \code{\link[MicrobiomeR]{object_handler}}.}

\item{method}{Choose an ordination method from 'PCoA', 'CCA', 'NMDS' or 'DPCoA', Default: 'PCoA'}

\item{distance}{Choose a distance method from 'bray', 'unifrac' or 'wunifrac', Default: 'wunifrac'}

\item{title}{The title of the plot, Default: NULL}

\item{save}{Save your plot, Default: FALSE}

\item{only_data}{Allows for only ordination data to be generated, Default: FALSE}
}
\value{
By default, it returns an ordination plot.
}
\description{
This function allows for ordination (which helps us to distinguish beta diversity relationships) to be plotted as well as for the corresponding data to be returned.
}
\examples{
\dontrun{
if (interactive()) {
  # An example ordination plot
  library(MicrobiomeR)
  data <- analyzed_silva
  plot <- ordination_plot(obj = data)
  plot
}
}
}
