% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{ordination_plot}
\alias{ordination_plot}
\title{Ordination Plot}
\usage{
ordination_plot(obj, method = "PCoA", distance = "wunifrac",
  color = "TreatmentGroup", title = NULL, only_data = FALSE)
}
\arguments{
\item{obj}{An object to be converted to a metacoder object with \code{\link[MicrobiomeR]{object_handler}}.}

\item{method}{Choose an ordination method from 'PCoA', 'CCA', 'NMDS' or 'DPCoA', Default: 'PCoA'}

\item{distance}{Choose a distance method from 'bray', 'unifrac' or 'wunifrac', Default: 'wunifrac'}

\item{color}{Choose the group or factor of which colors will be mapped to, Default: 'TreatmentGroup'}

\item{title}{The title of the plot, Default: NULL}

\item{only_data}{Allows for only ordination data to be generated, Default: FALSE}
}
\value{
By default, it returns an ordination plot.
}
\description{
This function allows for ordination (which helps us to distinguish beta diversity relationships) to be plotted as well as for the corresponding data to be returned.
}
\examples{
\dontrun{
if (interactive()) {
  # An example ordination plot
  library(MicrobiomeR)
  data <- analyzed_silva
  plot <- ordination_plot(obj = data)
  plot
}
}
}
\seealso{
View \code{\link{save_ordination_plots}} to save your ordination plot or multiple ordination plots.

Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_plot}},
  \code{\link{get_correlation_data}},
  \code{\link{get_correlation_plots}},
  \code{\link{get_heat_tree_parameters}},
  \code{\link{get_heat_tree_plots}},
  \code{\link{get_plot_limits}},
  \code{\link{melt_metacoder}}, \code{\link{save_barplot}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{save_ordination_plots}},
  \code{\link{top_coefficients_barplot}}
}
\concept{Visualizations}
