% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{otu_id_filter}
\alias{otu_id_filter}
\title{Filter OTU Ids from Metacoder Objects}
\usage{
otu_id_filter(obj, .f_transform = NULL, .f_filter = NULL,
  .f_condition = NULL, validated = FALSE, ...)
}
\arguments{
\item{obj}{A Taxmap/metacoder object.}

\item{.f_transform}{A function used for transforming the data.  Default: NULL}

\item{.f_filter}{A function used for summarising the data like 'sum' or 'mean'.  Default: NULL}

\item{.f_condition}{A function that takes the summarised data and applied a condition like x > 10000.  Default: NULL}

\item{validated}{This parameter provides a way to override validation steps.  Use carefully.  Default: FALSE}

\item{...}{An optional list of parameters to use in the .f_filter function specified}
}
\value{
Returns a metacoder object with otu_ids that pass the filters.
}
\description{
This function provides a flexible way to filter unwanted otu_ids from the taxmap object and from the
observations of a MicrobiomeR formatted object.
}
\details{
Get the otu_ids to keep by using purr and the user supplied transform and filter + condition formulas.
The purr packge allows the use of anonymous functions as described in the link below:

\url{https://jennybc.github.io/purrr-tutorial/ls03_map-function-syntax.html#anonymous_function,_formula}
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{validate_MicrobiomeR_format}},  \code{\link[MicrobiomeR]{transposer}},  \code{\link[MicrobiomeR]{transformer}}

\code{\link[taxa]{get_dataset}},  \code{\link[taxa]{filter_taxa}}

\code{\link[dplyr]{select_all}}

\code{\link[purrr]{map}},  \code{\link[purrr]{keep}}

Other Metacoder Filters: \code{\link{agglomerate_metacoder}},
  \code{\link{cov_filter}},
  \code{\link{otu_prevalence_filter}},
  \code{\link{otu_proportion_filter}},
  \code{\link{sample_filter}},
  \code{\link{taxa_prevalence_filter}},
  \code{\link{taxon_id_filter}}
}
\concept{Metacoder Filters}
