% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder.R
\name{otu_prevalence_filter}
\alias{otu_prevalence_filter}
\title{OTU Prevalence Filter (Metacoder)}
\usage{
otu_prevalence_filter(obj, minimum_abundance = 5,
  rel_sample_percentage = 0.5, validated = FALSE)
}
\arguments{
\item{obj}{A Taxmap/metacoder object.}

\item{minimum_abundance}{The minimum abundance needed per observation per sample.  Default: 5}

\item{rel_sample_percentage}{The percentage of samples per observation that meet the minimum abundance.  Default: 0.5}

\item{validated}{This parameter provides a way to override validation steps.  Use carefully.  Default: FALSE}
}
\value{
Returns a taxmap object that contains taxon_ids that have passed the above filter.
}
\description{
This function filters observations by thier prevelance across samples.
}
\details{
The otu_prevalence_filter filters taxon_ids that do not appear more than a certian amount of times (minimum abundance) in a certain percentage of
samples (rel_sample_percentage). The \href{http://web.stanford.edu/class/bios221/MicrobiomeWorkflowII.html#filtering}{phyloseq workflow} calls for a minimum abundance of 5 across %50 of the samples.
This filtering method is considered unsupervised, because it soley relies on the data in this experiment (OTU ids).
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{validate_MicrobiomeR_format}}
\code{\link[metacoder]{calc_prop_samples}}
\code{\link[dplyr]{filter}}
\code{\link[taxa]{filter_taxa}}

Other Metacoder Filters: \code{\link{agglomerate_metacoder}},
  \code{\link{cov_filter}}, \code{\link{otu_id_filter}},
  \code{\link{otu_proportion_filter}},
  \code{\link{sample_filter}},
  \code{\link{taxa_prevalence_filter}},
  \code{\link{taxon_id_filter}}
}
\concept{Metacoder Filters}
