% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq.R
\name{parse_taxonomy_greengenes2}
\alias{parse_taxonomy_greengenes2}
\title{@title Parse Taxonomy Greengenes2}
\usage{
parse_taxonomy_greengenes2(char.vec)
}
\description{
@description  A custom Greengenes parsing function
}
\details{
@param char.vec DESCRIPTION.

@return RETURN_DESCRIPTION
@export
}
