% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-plot.R
\name{plot_limits}
\alias{plot_limits}
\title{Get Plot Limits}
\usage{
plot_limits(x, y)
}
\arguments{
\item{x}{Data used to plot the x axis.}

\item{y}{Data used to plot the y axis.}
}
\value{
A vector that supplies the overarching x and y limits.
}
\description{
Get the limits of the plot using data values.
}
\seealso{
Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_data}},
  \code{\link{correlation_plots}},
  \code{\link{correlation_plot}},
  \code{\link{heat_tree_parameters}},
  \code{\link{heat_tree_plots}}, \code{\link{melt_taxmap}},
  \code{\link{ordination_plot}},
  \code{\link{save_barplot}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{save_ordination_plots}},
  \code{\link{top_coefficients_barplot}}
}
\concept{Visualizations}
