% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres_abund_filter.R
\name{pres_abund_filter}
\alias{pres_abund_filter}
\title{Filter phyloseq-object based on presence/absence}
\usage{
pres_abund_filter(ps, pres = 2, abund = 0.001, verbose = T)
}
\arguments{
\item{ps}{phyloseq-object}

\item{pres}{number of samples in which given ASV should at least be present to be included}

\item{abund}{abundance which given ASV should at least be present}

\item{verbose}{(logical) print information on number of ASVs included/excluded}
}
\value{
A filtered phyloseq-object
}
\description{
Filter phyloseq-object based on presence/absence
}
\examples{
data(esophagus)
phyloseq::ntaxa(esophagus)
esophagus_filtered <- esophagus \%>\% pres_abund_filter()
phyloseq::ntaxa(esophagus_filtered)
}
