% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{save_barplot}
\alias{save_barplot}
\title{Save Barplot}
\usage{
save_barplot(plot, filename)
}
\arguments{
\item{plot}{The plot object.}

\item{filename}{The name of the file. (an extension should not be included)}
}
\description{
Save a stacked barplot.
}
\examples{
\dontrun{
if (interactive()) {
  library(MicrobiomeR)
  save_barplot(plot = plot, filename = "stacked_bar_phylum")
}
}
}
\seealso{
Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_plot}},
  \code{\link{get_correlation_plots}},
  \code{\link{get_heat_tree_parameters}},
  \code{\link{get_heat_tree_plots}},
  \code{\link{get_plot_limits}},
  \code{\link{melt_metacoder}},
  \code{\link{ordination_plot}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{save_ordination_plots}},
  \code{\link{top_coefficients_barplot}}
}
\concept{Visualizations}
