% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{save_barplot}
\alias{save_barplot}
\title{Save Barplot}
\usage{
save_barplot(plot, filename)
}
\arguments{
\item{plot}{The plot object.}

\item{filename}{The name of the file. (an extension should not be included)}
}
\description{
Save a stacked barplot.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
