% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_plot.R
\name{save_correlation_plots}
\alias{save_correlation_plots}
\title{Save Correlation Plots}
\usage{
save_correlation_plots(corr, pairwise = FALSE, format = "tiff",
  start_path = "output", ...)
}
\arguments{
\item{corr}{A correlation plot list generated by correlation_plot or get_correlation_plots.}

\item{pairwise}{Denotes weather or not the plot list is pairwise.  Default: FALSE}

\item{format}{The format of the output image.  Default: 'tiff'}

\item{start_path}{The starting path of the output directory.  Default: 'output'}

\item{...}{An optional list of parameters to use in the get_output_dir function.}
}
\value{
An output directory that contains correlation plots.
}
\description{
This function saves correlation plots storred in a listlike object to an output folder.
}
\details{
This function creates an appropriate output directory, where it saves publication ready
plots.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{get_output_dir}}
\code{\link[ggplot2]{ggsave}}

Other Visualizations: \code{\link{correlation_plot}},
  \code{\link{get_correlation_plots}},
  \code{\link{get_plot_limits}}
}
\concept{Visualizations}
