% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-plot.R
\name{save_correlation_plots}
\alias{save_correlation_plots}
\title{Save Correlation Plots}
\usage{
save_correlation_plots(corr, format = "tiff", start_path = "output",
  ...)
}
\arguments{
\item{corr}{A correlation plot list generated by correlation_plot or correlation_plots.}

\item{format}{The format of the output image.  Default: 'tiff'}

\item{start_path}{The starting path of the output directory.  Default: 'output'}

\item{...}{An optional list of parameters to use in the output_dir function.}
}
\value{
An output directory that contains correlation plots.
}
\description{
This function saves correlation plots stored in a listlike object to an output folder.
}
\details{
This function creates an appropriate output directory, where it saves publication ready
plots.
}
\examples{
\dontrun{
if(interactive()){
# This example uses data that are no longer available in the MicrobiomeR package,
# however, they can be easily generated with \\code{\\link{MicrobiomeR}{as_analyzed_silva}}.
library(MicrobiomeR)
analyzed_silva <- as_MicrobiomeR_format(MicrobiomeR::raw_silva_2, "analyzed_format")
corr_plot <- correlation_plot(analyzed_silva, primary_rank = "Class", secondary_rank = "Phylum")
# Save to \\emph{./output/corr_plot} folder.
save_correlation_plots(corr_plot)
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{output_dir}}

\code{\link[ggplot2]{ggsave}}

Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_data}},
  \code{\link{correlation_plots}},
  \code{\link{correlation_plot}},
  \code{\link{heat_tree_parameters}},
  \code{\link{heat_tree_plots}},
  \code{\link{ordination_plots}},
  \code{\link{ordination_plot}}, \code{\link{plot_limits}},
  \code{\link{save_alpha_diversity_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{save_ordination_plots}},
  \code{\link{save_stacked_barplots}},
  \code{\link{stacked_barplots}},
  \code{\link{stacked_barplot}},
  \code{\link{top_coefficients_barplot}}
}
\concept{Visualizations}
