% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat-tree.R
\name{save_heat_tree_plots}
\alias{save_heat_tree_plots}
\title{Save Heat Tree Plots}
\usage{
save_heat_tree_plots(htrees, format = "tiff", start_path = "output",
  ...)
}
\arguments{
\item{htrees}{A named list of heat trees.}

\item{format}{The format of the output image.  Default: 'tiff'}

\item{start_path}{The starting path of the output directory.  Default: 'output'}

\item{...}{An optional list of parameters to use in the get_output_dir function.}
}
\value{
An output directory that contains heat tree plots.
}
\description{
This function saves heat tree plots storred in a list object to an output folder.
}
\details{
This function creates an appropriate output directory, where it saves publication ready
plots.
}
\examples{
\dontrun{
if(interactive()){
# This example uses data that are no longer available in the MicrobiomeR package,
# however, they can be easily generated with \\code{\\link{MicrobiomeR}{as_analyzed_format}}.
library(MicrobiomeR)
h_trees <- get_heat_tree_plots(MicrobiomeR::analyzed_silva, rank_list = c("Phylum", "Class"))
# Save to \\emph{./output/heat_trees} folder.
save_heat_tree_plots(h_trees)
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{get_output_dir}}
\code{\link[ggplot2]{ggsave}}

Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_plot}},
  \code{\link{get_correlation_plots}},
  \code{\link{get_heat_tree_parameters}},
  \code{\link{get_heat_tree_plots}},
  \code{\link{get_plot_limits}},
  \code{\link{save_barplot}},
  \code{\link{save_correlation_plots}},
  \code{\link{stacked_barplot}}
}
\concept{Visualizations}
