% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{save_ordination_plots}
\alias{save_ordination_plots}
\title{Save Ordination Plots}
\usage{
save_ordination_plots(ord, format = "tiff", start_path = "output", ...)
}
\arguments{
\item{ord}{An ordination plot list generated by get_ordination_plots.}

\item{format}{The format of the output image.  Default: 'tiff'}

\item{start_path}{The starting path of the output directory.  Default: 'output'}

\item{...}{An optional list of parameters to use in the get_output_dir function.}
}
\value{
An output directory that contains ordination plots.
}
\description{
This function saves ordination plots storred in a listlike object to an output folder.
}
\details{
This function creates an appropriate output directory, where it saves publication ready
plots.
}
\examples{
\dontrun{
if (interactive()) {
}
}
}
\seealso{
\code{\link[MicrobiomeR]{get_output_dir}}

\code{\link[ggplot2]{ggsave}}

Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_plot}},
  \code{\link{get_correlation_plots}},
  \code{\link{get_heat_tree_parameters}},
  \code{\link{get_heat_tree_plots}},
  \code{\link{get_plot_limits}},
  \code{\link{melt_metacoder}},
  \code{\link{ordination_plot}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{top_coefficients_barplot}}
}
\concept{Visualizations}
