% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination.R
\name{save_ordination_plots}
\alias{save_ordination_plots}
\title{Save Ordination Plots}
\usage{
save_ordination_plots(ord_plots, format = "tiff",
  start_path = "output", ...)
}
\arguments{
\item{ord_plots}{An ordination plot list generated by ordination_plots.}

\item{format}{The format of the output image.  Default: 'tiff'}

\item{start_path}{The starting path of the output directory.  Default: 'output'}

\item{...}{An optional list of parameters to use in the output_dir function.}
}
\value{
An output directory that contains ordination plots.
}
\description{
This function saves ordination plots stored in a listlike object to an output folder.
}
\details{
This function creates an appropriate output directory, where it saves publication ready
plots.
}
\examples{
\dontrun{
if (interactive()) {
}
}
}
\seealso{
\code{\link[MicrobiomeR]{output_dir}}

\code{\link[ggplot2]{ggsave}}

Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_data}},
  \code{\link{correlation_plots}},
  \code{\link{correlation_plot}},
  \code{\link{heat_tree_parameters}},
  \code{\link{heat_tree_plots}},
  \code{\link{ordination_plots}},
  \code{\link{ordination_plot}}, \code{\link{plot_limits}},
  \code{\link{save_alpha_diversity_plots}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{save_stacked_barplots}},
  \code{\link{stacked_barplots}},
  \code{\link{stacked_barplot}},
  \code{\link{top_coefficients_barplot}}
}
\concept{Visualizations}
