% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{save_stacked_barplots}
\alias{save_stacked_barplots}
\title{Save Stacked Barplots}
\usage{
save_stacked_barplots(sb_plots, format = "tiff", start_path = "output",
  ...)
}
\arguments{
\item{sb_plots}{A named list of stacked barplots.}

\item{format}{The format of the output image.  Default: 'tiff'}

\item{start_path}{The starting path of the output directory.  Default: 'output'}

\item{...}{An optional list of parameters to use in the output_dir function.}
}
\value{
An output directory that contains stacked barplot.
}
\description{
This function saves stacked barplot stored in a list object to an output folder.
}
\details{
This function creates an appropriate output directory, where it saves publication ready
plots.
}
\examples{
\dontrun{
if (interactive()) {
  # This example uses data that are no longer available in the MicrobiomeR package,
  # however, they can be easily generated with \\code{\\link{MicrobiomeR}{as_analyzed_format}}.
  library(MicrobiomeR)
  analyzed_silva <- as_MicrobiomeR_format(MicrobiomeR::raw_silva_2, "analyzed_format")
  alpha_div_plots <- alpha_diversity_plots(analyzed_silva)
  # Save to \\emph{./output/alpha_diversity} folder.
  save_alpha_diversity_plot(salpha_div_plots)
}
}
}
\seealso{
\code{\link[MicrobiomeR]{output_dir}}

\code{\link[ggplot2]{ggsave}}

Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_data}},
  \code{\link{correlation_plots}},
  \code{\link{correlation_plot}},
  \code{\link{heat_tree_parameters}},
  \code{\link{heat_tree_plots}},
  \code{\link{melt_metacoder}},
  \code{\link{ordination_plot}}, \code{\link{plot_limits}},
  \code{\link{save_alpha_diversity_plots}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{save_ordination_plots}},
  \code{\link{top_coefficients_barplot}}
}
\concept{Visualizations}
