% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{stacked_barplot}
\alias{stacked_barplot}
\title{Stacked Barplot}
\usage{
stacked_barplot(obj, tax_level = "Phylum", fill = "Phylum",
  xlabel = "Samples", title = NULL, palette_values)
}
\arguments{
\item{obj}{An object to be converted to a metacoder object with \code{\link[MicrobiomeR]{object_handler}}.}

\item{tax_level}{DESCRIPTION.}

\item{fill}{DESCRIPTION.}

\item{xlabel}{DESCRIPTION.}

\item{title}{The title or name of the plot.}

\item{palette_values}{DESCRIPTION.}
}
\value{
Returns a stacked barplot.
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Other Visualizations: \code{\link{correlation_plot}},
  \code{\link{get_correlation_plots}},
  \code{\link{get_heat_tree_parameters}},
  \code{\link{get_heat_tree_plots}},
  \code{\link{get_plot_limits}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}}
}
\concept{Visualizations}
