% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{melt_metacoder}
\alias{melt_metacoder}
\alias{convert_proportions}
\alias{stacked_barplot}
\title{Melt Metacoder Object}
\usage{
melt_metacoder(obj)

convert_proportions(melted_df, tax_level)

stacked_barplot(obj, tax_level = "Phylum", fill = "Phylum",
  xlabel = "Samples", faceted = FALSE, title = NULL, palette_values)
}
\arguments{
\item{obj}{An object to be converted to a metacoder object with \code{\link[MicrobiomeR]{object_handler}}.}

\item{melted_df}{A "melted" dataframe from the metacoder object's data.}

\item{tax_level}{The taxonomic level, Default: 'Phylum'}

\item{fill}{The taxonomic level by which the bars are filled, Default: 'Phylum'}

\item{xlabel}{The label of the x axis, Default: 'Samples'}

\item{faceted}{A boolean to determine if the barplot should be faceted by TreatmentGroup}

\item{title}{The title or name of the plot.}

\item{palette_values}{A list of the colors to input to be mapped to the plot palette.}
}
\value{
Returns a melted dataframe.

Returns a transformed dataframe.

Returns a stacked barplot.
}
\description{
Melt the metacoder or phyloseq tables into a dataframe.

Convert the dataframe abundance values to percent 100.

Create a stacked barplot to show relative abundance of taxa.
}
\examples{
\dontrun{
if (interactive()) {
  # An example stacked bar plot
  library(MicrobiomeR)
  data <- analyzed_silva
  palette <- get_color_palette(color_no = 12)
  plot <- stacked_barplot(obj = data, palette_values = palette)
  plot
}
}
}
\seealso{
Other Formatting: \code{\link{MicrobiomeR_Formats}},
  \code{\link{as_MicrobiomeR_format}},
  \code{\link{as_analyzed_format}},
  \code{\link{as_basic_format}},
  \code{\link{as_custom_format}},
  \code{\link{as_phyloseq_format}},
  \code{\link{as_raw_format}}

Other Data Manipulators: \code{\link{metacoder_comp_func_1}},
  \code{\link{transformer}}, \code{\link{transposer}},
  \code{\link{vlookup}}

Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_plot}},
  \code{\link{get_correlation_data}},
  \code{\link{get_correlation_plots}},
  \code{\link{get_heat_tree_parameters}},
  \code{\link{get_heat_tree_plots}},
  \code{\link{get_plot_limits}},
  \code{\link{ordination_plot}},
  \code{\link{save_barplot}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{save_ordination_plots}},
  \code{\link{top_coefficients_barplot}}
}
\concept{Data Manipulators}
\concept{Formatting}
\concept{Visualizations}
