% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{top_coefficients_barplot}
\alias{top_coefficients_barplot}
\title{Top Coefficients Barplot}
\usage{
top_coefficients_barplot(top_coefficients, title = NULL)
}
\arguments{
\item{top_coefficients}{PARAM_DESCRIPTION}
}
\value{
Returns a barplot of the top coefficients
}
\description{
A barplot to visualize which taxa have had the greatest impact on the overall gut composition.
}
\examples{
\dontrun{
if (interactive()) {
  library(MicrobiomeR)
  data <- analyzed_silva
  p <- permanova(obj = data)
  top_coefficients_barplot(p$top_coefficients)
}
}
}
\seealso{
View \code{\link{permanova}} to understand how to generate \code{top_coefficents}.

Other Visualizations: \code{\link{alpha_diversity_plot}},
  \code{\link{correlation_plot}},
  \code{\link{get_correlation_plots}},
  \code{\link{get_heat_tree_parameters}},
  \code{\link{get_heat_tree_plots}},
  \code{\link{get_plot_limits}},
  \code{\link{melt_metacoder}},
  \code{\link{ordination_plot}},
  \code{\link{save_correlation_plots}},
  \code{\link{save_heat_tree_plots}},
  \code{\link{save_ordination_plots}}
}
\concept{Visualizations}
