% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformer}
\alias{transformer}
\title{Transforming Tidy Data}
\usage{
transformer(
  .data,
  func,
  by = "column",
  ids = NULL,
  header_name = NULL,
  preserved_categories = TRUE,
  separated_categories = NULL,
  ...
)
}
\arguments{
\item{.data}{A matrix/data_frame/tibble for transforming.}

\item{func}{A function, which can be anonymous, that will be used to transform the data.  (e.g. proportions:
\emph{x/sum(x)}).}

\item{by}{This denotes how the data should be transformed (\strong{column/row}). Default: 'column'}

\item{ids}{The column to transpose by if \emph{by = 'row'}. Default: The first column.}

\item{header_name}{A name for the numeric data that will be transposed if \emph{by = 'row'}.}

\item{preserved_categories}{A logical denoting weather categorical data should be conserved if \emph{by = 'row'}.  A
value of FALSE will cause all categorical data except the \emph{ids} to be dropped.  A value of
TRUE will cause the categorical data to preserved by \emph{tidyr::unite}ing these columns.  Default: TRUE}

\item{separated_categories}{A vector containing ordered column names to use in a previously transposed
and categorically preserved table if \emph{by = 'row'}.  Retransposing with this set should yield an exact replicate of
the original data.  Default: NULL}

\item{...}{Additional arguments passed on to \emph{func}.}
}
\value{
A tibble that has been transformed.
}
\description{
This function transforms a table by rows or by columns.
}
\details{
This function transforms the supplied data using the \emph{func} parameter, which is used
in the purrr package.  The purr package allows the use of anonymous functions as described in the link below:

\url{https://jennybc.github.io/purrr-tutorial/ls03_map-function-syntax.html#anonymous_function,_formula}
}
\examples{
\dontrun{
if(interactive()){
# This example uses data that are no longer available in the MicrobiomeR package,
# however, they can be easily generated with \code{\link{MicrobiomeR}{as_basic_format}}.
 library(MicrobiomeR)
 basic_silva <- as_MicrobiomeR_format(MicrobiomeR::raw_silva_2, "basic_format")
 data <- basic_silva$data$taxa_abundance
 # Get proportions using the anonymous functions
 tax_props <- data \%>\% transformer(~./sum(.))
 # Get proportions using explicit functions
 alt_tax_props <- data \%>\% transformer(function(x)x/sum(x))
 }
}
}
\seealso{
\code{\link[MicrobiomeR]{transposer}}
\code{\link[dplyr]{select_all}},  \code{\link[dplyr]{select}}
\code{\link[purrr]{modify}}

Other Data Manipulators: 
\code{\link{convert_proportions}()},
\code{\link{metacoder_comp_func_1}()},
\code{\link{transposer}()},
\code{\link{vlookup}()}
}
\concept{Data Manipulators}
