% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transposer}
\alias{transposer}
\title{Transposing Tidy Data}
\usage{
transposer(.data, ids = NULL, header_name, preserved_categories = TRUE,
  separated_categories = NULL)
}
\arguments{
\item{.data}{A matrix/data_frame/tibble for transposing.}

\item{ids}{The column to transpose by.  Default: The first column.}

\item{header_name}{A name for the numeric data that will be transposed.}

\item{preserved_categories}{A logical denoting weather categorical data should be conserved.  A
value of FALSE will cause all categorical data except the \emph{ids} to be dropped.  A value of
TRUE will cause the categorical data to preserved by \emph{tidyr::unite}ing these columns.  Default: TRUE}

\item{separated_categories}{A vector containing ordered column names to use in a previously transposed
and categorically preserved table.  Retransposing with this set should yield an exact replicate of
the original data.  Default: NULL}
}
\value{
A transposed data table as a tibble.
}
\description{
This function transposes tables containing numeric and categorical data using the
tidyr package.
}
\details{
Transposing can help with preforming operations on the rows of your tibbles.
}
\examples{
\dontrun{
if(interactive()){
# This example uses data that are no longer available in the MicrobiomeR package,
# however, they can be easily generated with \\code{\\link{MicrobiomeR}{as_basic_format}}.
 library(MicrobiomeR)
 data <- MicrobiomeR::basic_silva$data$taxa_abundance
 trans_data <- data \%>\%
   transposer(ids = "taxon_id", header_name = "samples")
 retrans_data <- trans_data \%>\%
   transposer(ids="samples", header_name="taxon_id")
 }
}
}
\seealso{
\code{\link[tibble]{is_tibble}}

\code{\link[dplyr]{select_all}},  \code{\link[dplyr]{select}},  \code{\link[dplyr]{reexports}},  \code{\link[dplyr:tidyeval]{sym}},  \code{\link[dplyr:reexports]{one_of}}

\code{\link[tidyr]{gather}},  \code{\link[tidyr]{unite}},  \code{\link[tidyr]{spread}},  \code{\link[tidyr]{separate}}

\code{\link[stringr]{str_detect}},  \code{\link[stringr]{str_count}}
}
\concept{Data Manipulation}
