% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacoder-formatting.R
\name{which_format}
\alias{which_format}
\title{Which MicrobiomeR Format}
\usage{
which_format(obj)
}
\arguments{
\item{obj}{A Taxmap/metacoder object.}
}
\value{
If the format is verified it returns a character string denoting the identified format.
}
\description{
A function for looking at a metacoder object and returning the identified MicrobiomeR format.
}
\details{
This function is used to get basic information about the format of the taxmap object
that is supplied by the user.
}
\examples{
\dontrun{
if(interactive()){
library(MicrobiomeR)
fmt <- which fmt(MicrobiomeR::analyzed_silva)
print(fmt)
 }
}
}
\seealso{
Other Validation: \code{\link{is_analyzed_format}},
  \code{\link{is_basic_format}},
  \code{\link{is_phyloseq_format}},
  \code{\link{is_raw_format}},
  \code{\link{order_metacoder_data}},
  \code{\link{validate_MicrobiomeR_format}}
}
\concept{Validation}
